% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.set.cmp.R
\name{adjust.cmp}
\alias{adjust.cmp}
\alias{\S4method{adjust.cmp}{GGD}}
\title{Adjust cmp field}
\usage{
\S4method{adjust.cmp}{GGD}(this.mix.type = NULL,
         grad = c("default", "normal", "h", "v", "v2", "v3", "hv"))
}
\arguments{
\item{this.mix.type}{A numeric value to set \code{mix.type} field as an integer.

                     If \code{NULL} (the default), \code{mix.type} field will be set to
                     the smallest number which can represent the distribution model.
                     Normally, it should be an integer from \code{0} to \code{4}.
                     \code{NA} is allowed only if \code{cmp} field has no rows.}

\item{grad}{A character string indicating the method of gradation.

                     \code{"h"} for horizontal, \code{"v"} for vertical,
                     and \code{"hv"} for horizontal-vertical.
                     The number after \code{"v"} is the number of components.
                     Numberless \code{"v"} is an alias for \code{"v2"}.

                     \code{"normal"} is for a normal distribution.
                     \code{"default"} is for depending on \code{this.mix.type} argument.

                     If other than \code{"default"} is indicated,
                     \code{this.mix.type} argument is ignored.
                     If the current distribution model cannot be represented by
                     the indicated model, an error will occur.}
}
\value{
The adjusted \code{\link[ggd]{GGD}} object itself (invisible).
}
\description{
Simplifies the components in \code{cmp} field with retaining
the substance of the distribution. Also, you can make \code{cmp} field redundant inversely.
The \code{mix.type} field will be change accordingly.
}
\details{
\subsection{What this method can and cannot do}{
     This method may change the value of \code{mix.type} field and
     the number of rows in \code{cmp} field, but retains the substance of the distribution
     and does not change other fields.

     When \code{this.mix.type} is given, \code{mix.type} field will be set to the
     given value if the substance of the distribution will not change.
     For example, when \code{mix.type = 2} and \code{cmp} has 2 rows,
     you can indicate \code{this.mix.type = 4} to enhance \code{cmp} field into 4 rows
     with retaining the substance of the distribution.

     If you want to change the substance of the distribution,
     for example, if you want to change from a horizontal gradation to a vertical gradation,
     use \code{\link[ggd]{set.cmp}} instead.
 }
}
\examples{
 ## Usually the cmp field is simplified automatically,
 ## so you do not have to call adjust.cmp by your own.
 ## For example, in this case,
 ## the number of components is simplified to 1 automatically.
 a <- ggd.set.cmp( data.frame( mean = c( 0, 0 ), sd = c( 1.1, 1.1 ) ) )
 a$kind; a$mix.type; a$cmp   ## mix.type = 0: a Normal Distribution

 ## GGD$new() makes 2 components of normal distributions for convenience.
 ## You can use adjust.cmp in order to adjust the cmp field of a new object to 1 component.
 a <- GGD$new()
 a$kind; a$mix.type; a$cmp   ## mix.type = 2: Horizontal Gradational Distribution

 a$adjust.cmp()
 a$kind; a$mix.type; a$cmp   ## mix.type = 0: Normal Distribution

 ## You can also write as:
 a <- GGD$new()$adjust.cmp()
 a$kind; a$mix.type; a$cmp

 ## If you want to give redundancy to the components, you can also use adjust.cmp.
 ## Normal Distribution with 2 components.
 a$adjust.cmp( this.mix.type = 1 )
 a$kind; a$mix.type; a$cmp   ## mix.type = 1: Mean of 2 Normal Distributions

 ## Normal Distribution with 3 components.
 a$adjust.cmp( grad = "v3" )
 a$kind; a$mix.type; a$cmp   ## mix.type = 3: Vertical Gradational Distribution

 ## Normal Distribution with 4 components.
 a$adjust.cmp( grad = "hv" )
 a$kind; a$mix.type; a$cmp   ## mix.type = 4: H-V Gradational Distribution
}
\seealso{
\code{\link[ggd]{set.cmp}}
}
