% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{adjust.cmp.rownames}
\alias{adjust.cmp.rownames}
\alias{\S4method{adjust.cmp.rownames}{GGD}}
\title{Adjust each row name of cmp}
\usage{
\S4method{adjust.cmp.rownames}{GGD}()
}
\value{
The adjusted \code{\link[ggd]{GGD}} object itself (invisible).
}
\description{
Sets each row name of \code{cmp} field according to \code{mix.type} field.
Normally, users of this class do not need to call this method directly.
}
