% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.mean.sd.R
\name{adjust.median.mean.sd}
\alias{adjust.median.mean.sd}
\alias{\S4method{adjust.median.mean.sd}{GGD}}
\title{Adjust median, mean, and sd fields}
\usage{
\S4method{adjust.median.mean.sd}{GGD}()
}
\value{
The \code{\link[ggd]{GGD}} object itself (invisible).
}
\description{
Calculates the median, the mean, and the standard deviation of the distribution model and
then sets those values into the fields.
Before calling this method, you must set \code{cmp} and \code{mix.type} fields.
Normally, users of this class do not need to call this method directly.
}
