% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.apply.R
\name{apply.cmp.sub}
\alias{apply.cmp.sub}
\title{[Non-exported] Call applying function}
\usage{
apply.cmp.sub(f, values, obj)
}
\arguments{
\item{f}{A function to apply to elements in a column of \code{cmp} field.}

\item{values}{A vector of current values of the column of \code{cmp} field.}

\item{obj}{The \code{\link[ggd]{GGD}} object which calls this function.}
}
\value{
A vector of new values of the column of \code{cmp} field.
}
\description{
Calls a applying function to get new values of a column of \code{cmp} field.
}
\seealso{
\code{\link[ggd]{apply.cmp}}
}
