% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{apply.math2.cmp}
\alias{apply.math2.cmp}
\title{[Non-exported] Apply a Math2 group function}
\usage{
apply.math2.cmp(math2.f, dg.mean, dg.sd, obj)
}
\arguments{
\item{math2.f}{A \link[methods]{Math2} group function.}

\item{dg.mean}{A vector of integer for the 1st argument of \code{math2.f}
to be used for \code{mean} column.}

\item{dg.sd}{A vector of integer similar to \code{dg.mean}, for \code{sd} column.}

\item{obj}{The \code{\link[ggd]{GGD}} object to be processed.}
}
\value{
A processed \code{\link[ggd]{GGD}} object.
}
\description{
Applies a \link[methods]{Math2} group function
(\code{\link[base]{round}} or \code{\link[base]{signif}}) to the values of \code{cmp} field
of a \code{\link[ggd]{GGD}} object and adjusts other fields of the object accordingly.
}
\seealso{
\code{\link[ggd]{round.cmp}}
}
