% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmn.bisection.R
\name{bisection}
\alias{bisection}
\title{[Non-exported] Bisection method}
\usage{
bisection(f, interval, ftol = .Machine$double.eps * 16, xtol = ftol)
}
\arguments{
\item{f}{The destination function to solve.}

\item{interval}{A numeric vector of length 2 indicating the range of x value
on which the solution is to be searched.}

\item{ftol}{The function value tolerance.
When the absolute function value at either edge of the step is smaller
than \code{ftol}, the convergence is declared.}

\item{xtol}{The relative steplength tolerance.
When the relative steplength of x value is smaller than \code{xtol},
the convergence is declared.}
}
\value{
Solution of the equation f = 0 with tolerances.
}
\description{
Solves a equation via bisection method.
}
