% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection.sub}
\alias{bisection.sub}
\title{[Non-exported] Subfunction for bisection method}
\usage{
bisection.sub(f, interval, ftol, xtol)
}
\arguments{
\item{f}{The destination function to solve.}

\item{interval}{A vector of the range on which the solution is to be searched.}

\item{ftol}{The function value tolerance.
When the absolute function value at either edge of the step is smaller
than \code{ftol}, the convergence is declared.}

\item{xtol}{The relative steplength tolerance.
When the relative steplength of x value is smaller than \code{xtol},
the convergence is declared.}
}
\value{
Solution of the equation f = 0 with tolerances.
}
\description{
The entity that processes bisection method.
}
\seealso{
\code{\link[ggd]{bisection}}
}
