% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{calc.v.t4.via.integrate}
\alias{calc.v.t4.via.integrate}
\title{[Non-exported] Half variance computation for mix.type = 4}
\usage{
calc.v.t4.via.integrate(
  means,
  sds,
  mean = calc.mean(4, means, sds),
  get.lv = FALSE,
  get.uv = FALSE
)
}
\arguments{
\item{means}{The mean values of the normal distributions of the components.}

\item{sds}{The standard deviations of the normal distributions of the components.}

\item{mean}{The mean of the whole distribution.}

\item{get.lv}{If \code{TRUE}, computes the lower half variance.}

\item{get.uv}{If \code{TRUE}, computes the upper half variance.
Although it is not recommended to set
both \code{get.lv} and \code{get.uv} to \code{TRUE},
if both are \code{TRUE}, get.lv takes priority.
If both are \code{FALSE}, computes the whole variance.}
}
\value{
A list of the output of \code{\link[stats]{integrate}}
         as the result of the half/whole variance computation.
}
\description{
Using \code{\link[stats]{integrate}},
computes the lower/upper half variance for \code{mix.type = 4}.
Computing the whole variance with \code{\link[stats]{integrate}} is also enabled.
}
