% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmn.csv.R
\name{cat.table}
\alias{cat.table}
\title{[Non-exported] Output a table to a file}
\usage{
cat.table(x, file = "", top.text = "", append.str = "", digits = 15)
}
\arguments{
\item{x}{The object to be output, should be a matrix or data frame.}

\item{file}{A \link[base]{connection},
or a character string naming the file to print to.
If \code{""} (the default), it prints to the standard output connection,
the console unless redirected by \link[base]{sink}.}

\item{top.text}{A character string to be output as the first element of the header.}

\item{append.str}{A character string to be output after \code{x}.}

\item{digits}{The number of significant digits (see \link[base]{signif}) to print.
Integers from \code{1} to \code{22} with default \code{15} are allowed.}
}
\value{
An invisible \code{NULL}.
}
\description{
Outputs a data frame or a matrix to a file or \link[base]{connection} with the CSV format
using \link[base]{cat}.
The precision for real numbers can be indicated with \code{digits} argument.
}
\examples{
 ggd:::cat.table( data.frame( a = 1:6 * pi, b = pi^(1:6) ), top.text = "PI", digits = 10 )
}
