% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{cut.out.freq.edge}
\alias{cut.out.freq.edge}
\title{[Non-exported] Cut out small frequencies of edges}
\usage{
cut.out.freq.edge(x, freq)
}
\arguments{
\item{x}{A vector of x-coordinates of a frequency distribution.}

\item{freq}{A vector of frequencies of a frequency distribution.}
}
\value{
A list containing components
         \item{x}{
                 The vectors of x-coordinates after the cutoff.}
         \item{freq}{
                 The vectors of frequencies after the cutoff.}
}
\description{
Cuts out data with \code{0} or extremely small frequency on both edges of the range of
the x-coordinates from data of a frequency distribution.
}
