% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{d}
\alias{d}
\alias{\S4method{d}{GGD}}
\title{Probability density function}
\usage{
\S4method{d}{GGD}(x)
}
\arguments{
\item{x}{A vector of x-coordinates.}
}
\value{
The values of the probability density function for the given x-coordinates.
}
\description{
Gets the values of the probability density function for the given x-coordinates.
This method works like \code{\link[stats]{dnorm}} for a normal distribution.
}
\examples{
 a <- GGD$new()
 a$trace.q(
     data.frame( x = c( -0.67, 0, 0.53 ), p = c( 0.25, 0.5, 0.75 ) ),
     this.mix.type = 2 )
 a$d( c( -0.67, 0, 0.53 ) )
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )
}
\seealso{
\code{\link[ggd]{p}}, \code{\link[ggd]{q}}, \code{\link[ggd]{r}}
}
