% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{dp.t3}
\alias{dp.t3}
\title{[Non-exported] PDF/CDF for mix.type = 3}
\usage{
dp.t3(x, means, sds, f.t3)
}
\arguments{
\item{x}{A vector of x-coordinates.}

\item{means}{The vector of mean values of the 3 components.}

\item{sds}{The vector of sd values of the 3 components.}

\item{f.t3}{A function handle,
\code{ggd:::f.t3.d} for PDF or \code{ggd:::f.t3.p} for CDF.}
}
\value{
The vector of values of the probability density function
         or the cumulative distribution function.
}
\description{
Calculates the values of the probability density function or
the cumulative distribution function of the GGD model with \code{mix.type = 3}.
Both \code{means} and \code{sds} vectors need 3 elements for this function;
so where with two components, you must set \code{means[3]} and \code{sds[3]}
the same values of \code{means[1]} and \code{sds[1]}.
}
