% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmn.data.frame.R
\name{extract.complete.x.y}
\alias{extract.complete.x.y}
\title{[Non-exported] Extract complete-case data in two columns}
\usage{
extract.complete.x.y(data, x, y, y.name)
}
\arguments{
\item{data}{A data frame which represents a relation of x- and y-coordinates.
It should contain 2 columns for \code{x} and \code{y}.}

\item{x}{The column name or column number indicating \code{x} of \code{data},
the column of x-coordinates.}

\item{y}{The column name or column number indicating \code{y} of \code{data},
the column of y-coordinates.}

\item{y.name}{The column name of \code{y} column of the new data frame.}
}
\value{
A data frame with 2 columns named \code{"x"} and \code{y.name}.
}
\description{
Extracts complete-case data from the indicated data frame and makes a new data frame
with 2 columns named \code{"x"} and \code{y.name}.
}
