% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{extract.freq.data}
\alias{extract.freq.data}
\title{[Non-exported] Extract complete-case frequency distribution data}
\usage{
extract.freq.data(data, x, freq)
}
\arguments{
\item{data}{A data frame which represents a relation of x-coordinates and frequencies.
It should contain 2 columns for \code{x} and \code{freq}.}

\item{x}{The column name or column number indicating \code{x} of \code{data},
the column of x-coordinates.}

\item{freq}{The column name or column number indicating \code{freq} of \code{data},
the column of frequencies.}
}
\value{
A data frame with 2 columns named \code{"x"} and \code{"freq"}.
}
\description{
Checks errors of the indicated frequency distribution data frame and makes
a complete-case data frame with 2 columns named \code{"x"} and \code{"freq"}.
}
\seealso{
\code{\link[ggd]{extract.complete.x.y}}
}
