% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{get.cmp.with.nls.coef}
\alias{get.cmp.with.nls.coef}
\title{[Non-exported] Making cmp field with the result of nls}
\usage{
get.cmp.with.nls.coef(coefs, mix.type, grad, ncmp, eq.mean, eq.sd)
}
\arguments{
\item{coefs}{The numeric vector got by \code{\link[stats]{coef}}
with the result of \code{\link[stats]{nls}}.}

\item{mix.type}{The value for \code{mix.type} field.}

\item{grad}{A character string indicating the method of gradation.

                     In this function, \code{grad} argument has no priority over
                     \code{mix.type} argument.
                     Therefore, \code{grad = "v2"} or \code{"v3"} works
                     only if \code{mix.type = 3} is indicated.
                     And if \code{mix.type = 3} is indicated,
                     either \code{grad = "v2"} or \code{"v3"} must be indicated.
                     If \code{"v3"}, the number of components will forced to be 3,
                     even if the values in \code{coefs} represent a 2-component model.}

\item{ncmp}{Number of components for the custom distribution.}

\item{eq.mean}{A logical; the flag whether to make all of the mean values of
the normal distributions of the components to be equal.}

\item{eq.sd}{A logical; the flag whether to make all of the standard deviations of
the normal distributions of the components to be equal.}
}
\value{
A data frame for \code{cmp} field
         according to the result of \code{\link[stats]{nls}}.
}
\description{
Makes a data frame for \code{cmp} field of a \code{\link[ggd]{GGD}} object
with the result of \code{\link[stats]{nls}}.
}
