% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{get.d.freq}
\alias{get.d.freq}
\title{[Non-exported] Probability density values from a frequency distribution}
\usage{
get.d.freq(x, freq, total = sum(freq))
}
\arguments{
\item{x}{A vector of x-coordinates. Duplicate values are not allowed.}

\item{freq}{A vector of frequencies following \code{x}.
The values of frequencies must be positive.
Both integers and real numbers are allowed for the values.}

\item{total}{Total value of the frequencies.

                 If \code{NULL} (the default),
                 the total of \code{freq}, i.e., \code{\link[base]{sum}(data[[freq]])}
                 (on \code{\link[stats]{complete.cases}} of \code{x} and \code{freq})
                 is used for it.}
}
\value{
The vector of expected probability density value at each of x-coordinates.
}
\description{
Gets the probability density value at each of the x-coordinates based on
a frequency distribution.
}
\examples{
 x <- seq( -2, 2, 0.2 )
 freq <- c( 1517,  2292,  2513,  2763,  3724,  4046,  4713,
            7947, 10997, 11824, 11133,  7868,  4692,  4103,
            3698,  2740,  2549,  2284,  1499,  1147,   918 )
 ggd:::get.d.freq( x, freq )
 plot( x, ggd:::get.d.freq( x, freq ) )
}
