% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{get.nls.params}
\alias{get.nls.params}
\title{[Non-exported] Parameters for nls}
\usage{
get.nls.params(
  x,
  freq,
  total,
  mix.type,
  grad,
  ncmp,
  eq.mean,
  eq.sd,
  start.level
)
}
\arguments{
\item{x}{A vector of x-coordinates arranged in ascending order
and non-duplicated.}

\item{freq}{A vector of frequencies following \code{x}.}

\item{total}{Total value of the frequencies.}

\item{mix.type}{The value for \code{mix.type} field.}

\item{grad}{A character string indicating the method of gradation.

                     In this function, \code{grad} argument have no priority
                     over \code{mix.type} argument.
                     Therefore, \code{grad = "v2"} or \code{"v3"} works only if
                     \code{mix.type = 3} is indicated.
                     And if \code{mix.type = 3} is indicated,
                     either \code{grad = "v2"} or \code{"v3"} must be indicated.
                     If \code{"v3"}, the number of components will forced to be 3.}

\item{ncmp}{Number of components for the custom distribution.}

\item{eq.mean}{A logical; the flag whether to make all of the mean values of
the normal distributions of the components to be equal.}

\item{eq.sd}{A logical; the flag whether to make all of the standard deviations of
the normal distributions of the components to be equal.}

\item{start.level}{An integer from \code{0} to \code{3} or \code{NA};
                     the desired level at which to guess
                     the initial \code{start} parameters for \code{\link[stats]{nls}}.

         The detail of each level is:
         \itemize{
             \item from \code{0} to \code{3}:
                     Same as \code{start.level} in arguments of \code{\link[ggd]{nls.freq}}.
             \item \code{NA}: All elements in \code{start} will be \code{NA}.
         }}
}
\value{
A list containing \code{formula} and \code{start} for \code{\link[stats]{nls}}
         and adopted \code{start.level}.
         The value of \code{start.level} may be different from the argument value
         if initial guessing has failed at the desired level.
}
\description{
Gets a list of main arguments for \code{\link[stats]{nls}} (except for \code{data}),
based on the configuration arguments.
}
