% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{get.p.freq}
\alias{get.p.freq}
\title{[Non-exported] Probabilities of frequency distribution}
\usage{
get.p.freq(freq, total)
}
\arguments{
\item{freq}{A vector of frequencies.
The x-coordinates corresponding to these frequencies are expected to be
at the center of the cells of the frequency distribution.}

\item{total}{Total value of the frequencies.}
}
\value{
A vector of probabilities.
}
\description{
Gets a vector of probabilities that the random variable is less than or equal to the value of
the x-coordinate of the frequency distribution.
This function assumes that each x-coordinate of the frequency distribution is at the center
of a cell and that an equal number of samples of both upper and lower outside of the range
have been excluded.
}
