% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\docType{class}
\name{GGD-class}
\alias{GGD-class}
\alias{GGD}
\title{Gradational Gaussian Distribution class}
\description{
The class provides the Gradational Gaussian Distribution.
}
\details{
\subsection{Overview about GGD}{
     The \bold{gradational Gaussian distribution (GGD)} is one of
     continuous distribution models for mainly modeling asymmetric unimodal data
     which do not follow a normal distribution.

     The GGD is alike the Gaussian mixture model (GMM) but different.
     The GMM is represented by linear combinations of some normal distributions,
     and is often used for clustering of mixed data.
     On the other hand, the GGD is a distribution model of
     which mixes some normal distributions with gradually changing ratio
     along the x-axis or y-axis directions,
     and it treats data which do not follow a normal distribution as they are.

     About the \bold{horizontal gradational Gaussian distribution}
     \eqn{\mathcal{G}[\mathcal{N}_1 \rightarrow \mathcal{N}_2]},
     it is expressed as
     \deqn{\mathcal{G}[\mathcal{N}_1 \rightarrow \mathcal{N}_2] =
           h_1(x) \ \mathcal{N}_1 + h_2(x) \ \mathcal{N}_2,}
     where \eqn{h_1} and \eqn{h_2} are the mixing ratio functions defined
     using \eqn{\Phi_1} and \eqn{\Phi_2}, the cumulative distribution functions of
     \eqn{\mathcal{N}_1} and \eqn{\mathcal{N}_2} as
     \deqn{h_1(x) = 1 - \Phi_1(x), \\
           h_2(x) = \Phi_2(x). \quad \ \ \ }

     Therefore, \eqn{h_1(x)} decreases gradually along x-axis as
     \eqn{x:-\infty \to \infty \ \Rightarrow \ h_1(x):1 \to 0},
     and \eqn{h_2(x)} increases gradually as
     \eqn{x:-\infty \to \infty \ \Rightarrow \ h_2(x):0 \to 1}.

     Here, \eqn{\mathcal{N}_1} is dominant on the left (lower) side of x-axis,
     and \eqn{\mathcal{N}_2} is on the right (upper) side.
     So we call \eqn{\mathcal{N}_1} the \bold{left- (lower-) side} distribution,
     and \eqn{\mathcal{N}_2} the \bold{right- (upper-) side} distribution.

     About the \bold{vertical gradational Gaussian distribution}
     \eqn{\mathcal{G}[\mathcal{N}_1 \uparrow \mathcal{N}_2]},
     it is expressed as
     \deqn{\mathcal{G}[\mathcal{N}_1 \uparrow \mathcal{N}_2] =
           v_1(x) \ \mathcal{N}_1 + v_2(x) \ \mathcal{N}_2,}

     where the mixing ratio functions \eqn{v_1} and \eqn{v_2} are defined
     using \eqn{f_1} and \eqn{f_2}, the probability density functions of
     \eqn{\mathcal{N}_1} and \eqn{\mathcal{N}_2} as
     \deqn{v_1(x) = 1 - \dfrac{f_1(x)}{f_1(\mu_1)}, \\
           v_2(x) = \dfrac{f_2(x)}{f_2(\mu_2)}, \quad \ \ \ }
     where \eqn{\mu_1} and \eqn{\mu_2} are the mean values of
     \eqn{\mathcal{N}_1} and \eqn{\mathcal{N}_2}, respectively.

     Here, \eqn{v_1(x)} is decreasing gradually along y-axis,
     i.e. value of \eqn{f_1(x)} as
     \deqn{f_1(x):0 \to \max_{x \in (-\infty, \infty)}f_1(x) \ \Rightarrow \ v_1(x):1 \to 0,}

     and inversely, \eqn{v_2} is increasing gradually as
     \deqn{f_2(x):0 \to \max_{x \in (-\infty, \infty)}f_2(x) \ \Rightarrow \ v_2(x):0 \to 1.}

     Or, using \eqn{\mu_1} and \eqn{\mu_2},
     the mean values of \eqn{\mathcal{N}_1} and \eqn{\mathcal{N}_2}, we can write as
     \deqn{x:-\infty \to \mu_1 \ \Rightarrow \ v_1(x):1 \to 0, \ \ \ \
           x:\mu_1 \to \infty \ \Rightarrow \ v_1(x):0 \to 1, \\
           x:-\infty \to \mu_2 \ \Rightarrow \ v_2(x):0 \to 1, \ \ \ \
           x:\mu_2 \to \infty \ \Rightarrow \ v_2(x):1 \to 0.}

     Here, we call \eqn{\mathcal{N}_1} the \bold{tail-side} distribution,
     and \eqn{\mathcal{N}_2} the \bold{top-side} distribution.
     Normally, the shape of the probability density function of
     \eqn{\mathcal{G}[\mathcal{N}_1 \uparrow \mathcal{N}_2]} should be gradually
     change from the shape of \eqn{f_1} to \eqn{f_2} along y-axis,
     and the top of it should be same as or nearby the top of \eqn{f_2}.

     However, note that \eqn{\mu_1} and \eqn{\mu_2} are not necessarily equal to each other.
     Therefore, depending on how \eqn{\mathcal{N}_1} and \eqn{\mathcal{N}_2} are taken,
     The tops of \eqn{f_1} and \eqn{f_2} could be far apart from each other,
     and moreover, the top of \eqn{\mathcal{G}[\mathcal{N}_1 \uparrow \mathcal{N}_2]}
     could be nearby the top of \eqn{f_1}, instead of \eqn{f_2}.
     That may be contrary to the intuitive image of the "vertical gradational distribution",
     but it is not prohibited.

     About the \bold{3-component vertical gradational Gaussian distribution},
     you can divide the tail-side distribution along x-axis into left (lower) side
     and right (upper) side.

     In this case, we write a sign of the distribution as
     \eqn{\mathcal{G}[\mathcal{N}_1 \uparrow \mathcal{N}_2 \downarrow \mathcal{N}_3]}.
     Generally it is defined as
     \deqn{\mathcal{G}[\mathcal{N}_1 \uparrow \mathcal{N}_2 \downarrow \mathcal{N}_3] =
           v_1(x) \ \mathcal{N}_1 + v_2(x) \ \mathcal{N}_2 + v_3(x) \ \mathcal{N}_3,}

     where the mixing ratio functions \eqn{v_1} and \eqn{v_3} are defined as
     \deqn{v_1(x) =
           \begin{cases}
           1 - \dfrac{f_1(x)}{f_1(\mu_1)} & (x \leq \mu_1), \\
           0 & (x > \mu_1),
           \end{cases} \\
           v_3(x) =
           \begin{cases}
           0 & (x < \mu_3), \\
           1 - \dfrac{f_3(x)}{f_3(\mu_3)} & (x \geq \mu_3),
           \end{cases}}

     and \eqn{v_2(x)} for the top side is defined as same as with 2 components.
     Here, \eqn{\mu_1} and \eqn{\mu_3} are the mean values of
     \eqn{\mathcal{N}_1} and  \eqn{\mathcal{N}_3}, respectively.
     We call \eqn{\mathcal{N}_1} the \bold{left- (lower-) tail-side} distribution,
     and \eqn{\mathcal{N}_3} the \bold{right- (upper-) tail-side} distribution,
     regardless whether \eqn{\mu_1} and \eqn{\mu_3} is greater.

     About the \bold{horizontal-vertical gradational Gaussian distribution}
     \eqn{\mathcal{G}[\mathcal{G}_1 \rightarrow \mathcal{G}_2]},
     2 vertical GGDs \eqn{\mathcal{G}_1} and \eqn{\mathcal{G}_2} can mixture
     as same as 2 normal distributions \eqn{\mathcal{N}_1} and \eqn{\mathcal{N}_2}
     for the horizontal GGD. It is expressed as
     \deqn{\ \ \ \ \mathcal{G}[\mathcal{G}_1 \rightarrow \mathcal{G}_2] =
           h_1(x) \ \mathcal{G}_1 + h_2(x) \ \mathcal{G}_2,}
     \deqn{\mathcal{G}_1 =
                 \mathcal{G}[\mathcal{N}_{1,1} \uparrow \mathcal{N}_{1,2}] =
                 v_{1,1}(x) \ \mathcal{N}_{1,1} + v_{1,2}(x) \ \mathcal{N}_{2,1},}
     \deqn{\mathcal{G}_2 =
                 \mathcal{G}[\mathcal{N}_{2,1} \uparrow \mathcal{N}_{2,2}] =
                 v_{2,1}(x) \ \mathcal{N}_{2,1} + v_{2,2}(x) \ \mathcal{N}_{2,2}.}

     These are the kinds of distribution models supported by this package.
 }
 \subsection{The probability density function and the cumulative distribution function}{
     In \code{GGD} class, \code{mix.type} is very important field that determines
     how the normal distributions of the components (shown in \code{cmd}) are mixed.
     In other words, \code{mix.type} determines the format of
     the probability density function and the cumulative distribution function.

     The probability density function \eqn{g(x)}
     and the cumulative distribution function \eqn{\Psi(x)}
     of each distribution model are defined as follows:

     \describe{
         \item{\code{mix.type = 0} (normal distribution)}{\deqn{
             g(x) = f_1(x) \\
             \Psi(x) = \Phi_1(x) }}

         \item{\code{mix.type = 1} (mean of 2 normal distributions;
                             a Gaussian mixture model (GMM), not a GGD)}{\deqn{
             g(x) = \dfrac{f_1(x) + f_2(x)}{2} \ \\
             \Psi(x) = \dfrac{\Phi_1(x) + \Phi_2(x)}{2} }}

         \item{\code{mix.type = 2} (horizontal GGD)}{\deqn{
             g(x) = \left( 1 - \Phi_1(x) \right) f_1(x)
                     + \Phi_2(x) f_2(x) \\
             \ \\
             \Psi(x) = \Phi_1(x) - \dfrac{\Phi_1(x)^2}{2} +
                         \dfrac{\Phi_2(x)^2}{2} \qquad \ \ }}

         \item{\code{mix.type = 3}, \code{grad = "v2"} (2-component vertical GGD)}{\deqn{
             g(x) = \left( 1 - \dfrac{f_1(x)}{f_1(\mu_1)} \right) f_1(x) +
                    \dfrac{f_2(x)^2}{f_2(\mu_2)} \\
             \ \\
             \Psi(x) = \Phi_1(x) - \dfrac{\Phi^*_1(x)}{\sqrt{2}} +
                       \dfrac{\Phi^*_2(x)}{\sqrt{2}} \qquad \; \: }}

         \item{\code{mix.type = 3}, \code{grad = "v3"} (3-component vertical GGD)}{\deqn{
             g(x) = g_1(x) + g_2(x) + g_3(x) \hspace{14.5em} \\
             \ \\
             g_1(x) = \left( 1 - \dfrac{f_1(x)}{f_1(\mu_1)} \right) f_1(x)
                      \quad (x \leq \mu_1; \ \mathrm{otherwise} \ 0) \\
             \ \\
             g_2(x) = \dfrac{f_2(x)^2}{f_2(\mu_2)} \hspace{15.6em} \\
             \ \\
             g_3(x) = \left( 1 - \dfrac{f_3(x)}{f_3(\mu_3)} \right) f_3(x)
                      \quad (x \geq \mu_3; \ \mathrm{otherwise} \, 0) \\
             \ \\
             \Psi(x) = \min \left( \Phi_1(x) - \dfrac{\Phi^*_1(x)}{\sqrt{2}},
                                 \ \dfrac{2 - \sqrt{2}}{4} \right) \hspace{10em} \\
                       + \dfrac{\Phi^*_2(x)}{\sqrt{2}}
                       + \max \left( 0, \ \Phi_3(x) - \dfrac{\Phi^*_3(x)}{\sqrt{2}} -
                                     \dfrac{2 - \sqrt{2}}{4} \right) }}

         \item{\code{mix.type = 4} (horizontal-vertical GGD)}{\deqn{
             g(x) = \left( 1 - \Psi_1(x)\ \right) g_1(x) + \Psi_2(x) g_2(x)
                    \hspace{6em} \ \ \\
             \ \\
             g_i(x) = \left( 1 - \dfrac{f_{i,1}(x)}{f_{i,1}(\mu_{i,1})} \right)
                      f_{i,1}(x) + \dfrac{f_{i,2}(x)^2}{f_{i,2}(\mu_{i,2})} \\
             \ \\
             \Psi(x) = \Psi_1(x) - \dfrac{\Psi_1(x)^2}{2} +
                                   \dfrac{\Psi_2(x)^2}{2} \hspace{9em} \\
             \ \\
             \Psi_i(x) = \Phi_{i,1}(x) - \dfrac{\Phi^*_{i,1}(x)}{\sqrt{2}} +
                         \dfrac{\Phi^*_{i,2}(x)}{\sqrt{2}} \hspace{3em} \ }}
     }

     Where \eqn{f_i} is the probability density function of
     the normal distribution \eqn{\mathcal{N}_i},
     \eqn{\Phi_i} and \eqn{\Phi_{i,j}} are the cumulative distribution functions of
     \eqn{\mathcal{N}_i} and \eqn{\mathcal{N}_{i,j}},
     and \eqn{\mu_i} is the mean value of \eqn{\mathcal{N}_i},
     and then \eqn{\Phi^*_i} and \eqn{\Phi^*_{i,j}} are the cumulative distribution function
     of the normal distributions \eqn{\mathcal{N}(\mu_i, (\sigma_i / \sqrt{2})^2)} and
     \eqn{\mathcal{N}(\mu_{i,j}, (\sigma_{i,j} / \sqrt{2})^2)}, respectively.

     For each of all above GGD models,
     we can say that \eqn{\Psi(x)} is a \eqn{0 \to 1} monotonical increasing function,
     regardless of how the normal distributions \eqn{\mathcal{N}_i} or
     \eqn{\mathcal{N}_{i,j}} are taken.
     This is because, \eqn{g(x)} is the derivative of \eqn{\Psi(x)} and it has
     \eqn{g(x) \geq 0} for \eqn{\forall x \in (-\infty, \infty)},
     moreover \eqn{\Psi(-\infty) = 0}, and \eqn{\Psi(\infty) = 1}.
 }
}
\section{Fields}{

\describe{
\item{\code{kind.index}}{An integer; the index number of the kind of the distribution model.}

\item{\code{kind}}{A string; the name of the kind of the distribution model.}

\item{\code{mix.type}}{An integer which represents how to mix normal distributions
                         of the components.

                         The type of the distribution model and the number of rows
                         in \code{cmp} field will be as follows with this value:
                         \itemize{
                             \item \code{0} : Normal distribution.
                                       \code{cmp} has only 1 row.
                             \item \code{1} : Mean of 2 normal distributions.
                                       \code{cmp} has 2 rows.
                             \item \code{2} : Horizontal gradational distribution.
                                       \code{cmp} has 2 rows.
                             \item \code{3} : Vertical gradational distribution.
                                       \code{cmp} has 2 or 3 rows.
                             \item \code{4} : Horizontal-vertical gradational distribution.
                                              \code{cmp} has 4 rows.
                             \item \code{5} : Customized distribution.
                                              Any distribution can be represented
                                              by a user-defined density function
                                              in \code{custom.d} field
                                              (allowed even not representing
                                               a probability distribution,
                                               but should be integrable).
                                              The number of rows in \code{cmp} is unlimited.
                         }

                         With \code{mix.type = 1}, the distribution model is not
                         a gradational Gaussian distribution (GGD), but a kind of
                         Gaussian mixture model (GMM).
                         This is provided for comparing GGD with GMM.}

\item{\code{cmp}}{A data frame with 2 numeric columns which shows
                         the normal distributions of the components.

                         \code{mean} column represents the mean values of the components,
                         and \code{sd} column represents the standard deviations.

                         Where \code{mix.type} is in \code{0:3},
                         it has 1 to 3 rows named like "\code{n.i}".
                         Where \code{mix.type = 4},
                         it has 4 rows named like "\code{n.i.j}".}

\item{\code{custom.d}}{The density function defined by user for the custom distribution.

             The function for \code{custom.d} must receive 2 arguments and
             return a vector of numeric values as the densities.
             The 2 arguments are
             \itemize{
                 \item   A vector of numeric values for the x-coordinates.
                         \code{custom.d} must return a vector of same length
                         as this argument.
                 \item   A data frame of the mean values and standard deviations
                         of the components. That is, the \code{cmp} field.
             }

             The default is \code{function(x, cmp) dnorm(x, cmp$mean[1], cmp$sd[1])}.
             However, the default can be changed without notice,
             so you should not expect it to be a probability density function
             of a normal distribution.}

\item{\code{custom.p}}{The cumulative distribution function defined by user.

             As with \code{custom.d},
             the function for \code{custom.p} must receive 2 arguments and
             return a numeric value of the cumulative distribution function,
             i.e. the probability of that a value of the random variable is
             less than or equal to the given x-coordinate.

             The 2 arguments are similar to \code{custom.d},
             but the length of the first argument is always \code{1}.

             The default is
             \code{function(x, cmp) integrate(function(x) custom.d(x, cmp), -Inf, x)$value}.
             It will work but may be slow and not very accurate.
             Therefore, this field is provided so that you can define your own function
             that will be faster or more accurate.

             This field is basically for custom distributions,
             but you can also use it for other supported models if you want.
             See \code{\link[ggd]{p}} or \code{\link[ggd]{q}} method.}

\item{\code{median}}{A numeric; the median of the distribution.}

\item{\code{mean}}{A numeric; the mean of the distribution.}

\item{\code{sd}}{A numeric; the standard deviation of the distribution.}

\item{\code{lsd}}{A numeric; the semi-standard deviation lower than mean.}

\item{\code{usd}}{A numeric; the semi-standard deviation upper than mean.}

\item{\code{mean.abs.error}}{A numeric;
the estimated modulus of the absolute error for \code{mean}.}

\item{\code{sd.abs.error}}{A numeric;
the estimated modulus of the absolute error for \code{sd}.}

\item{\code{lsd.abs.error}}{A numeric;
the estimated modulus of the absolute error for \code{lsd}.}

\item{\code{usd.abs.error}}{A numeric;
                         the estimated modulus of the absolute error for \code{usd}.

                         Where \code{mix.type = 4}, to compute the semi-standard deviations,
                         \code{\link[stats]{integrate}} function is used.
                         And the modulus of the absolute errors which
                         \code{\link[stats]{integrate}} function has reported
                         will be set into these \code{*.abs.error} fields.

                         Where \code{mix.type = 5}, \code{mean} and \code{sd} fields
                         are also computed with \code{\link[stats]{integrate}} function.
                         And then the modulus of the absolute errors which
                         \code{\link[stats]{integrate}} function has reported
                         will be set into \code{*.abs.error} fields.
                         In addition, if a discrete function is given for \code{custom.d},
                         these values will not be calculated correctly.}
}}


\seealso{
\code{\link[ggd]{set.cmp}},
         \code{\link[ggd]{nls.freq}}, \code{\link[ggd]{ggd.nls.freq.all}},
         \code{\link[ggd]{trace.q}},
         \href{https://github.com/Kimitsuna-Goblin/GGD}{README.md} (GitHub)
}
