% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{ggd.cor.vs.freq}
\alias{ggd.cor.vs.freq}
\title{Correlation between GGD and a frequency distribution}
\usage{
ggd.cor.vs.freq(objs, x, freq, total = sum(freq), cor.method = NULL)
}
\arguments{
\item{objs}{A \code{\link[ggd]{GGD}} object or a list of
\code{\link[ggd]{GGD}} objects.}

\item{x}{The vector of the x-coordinates of the frequency distribution.}

\item{freq}{The vector of the frequencies of the frequency distribution.}

\item{total}{Total value of the frequencies.}

\item{cor.method}{The \code{method} argument for \code{\link[stats]{cor}}.
It represents the correlation coefficient method.
If \code{NULL}, it uses the default method of \code{\link[stats]{cor}}.
See \code{\link[stats]{cor}} for more information.}
}
\value{
A vector of correlation coefficients.
         Its order follows the order of the elements of the \code{objs} argument.
         If \code{cmp} field of an \code{\link[ggd]{GGD}} object in \code{objs} has no rows,
         \code{NA} will be set to the element.
}
\description{
Computes correlation coefficients between Gradational Gaussian Distributions and
a frequency distribution.
Note, in this function, the arguments to give the frequency distribution are
two numeric vectors like as other non-exported functions, not a data frame.
}
\examples{
 df <- data.frame( x     = seq( -2, 2, 0.2 ),
                   freq  = c( 1517,  2292,  2513,  2763,  3724,  4046,  4713,
                              7947, 10997, 11824, 11133,  7868,  4692,  4103,
                              3698,  2740,  2549,  2284,  1499,  1147,   918 ) )
 a <- ggd.nls.freq( df, kind = "Mean-Eq.*Vertical" )$obj
 ggd.cor.vs.freq( a, df$x, df$freq )

 objs <- list( ggd.nls.freq( df, kind = "Mean-Eq.*Vertical" )$obj,
               ggd.nls.freq( df, kind = "Sigma-Eq.*Vertical" )$obj,
               ggd.nls.freq( df, kind = "Mean-Diff.*Sigma-Diff.*Vertical" )$obj )
 ggd.cor.vs.freq( objs, df$x, df$freq )
}
