% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{ggd.kind.index}
\alias{ggd.kind.index}
\title{Index number of the kind of distribution}
\usage{
ggd.kind.index(objs, undef.err = FALSE)
}
\arguments{
\item{objs}{A vector or a list of elements indicating the kind of distribution.
                     Each element must be a character string of a regular expression pattern
                     matching to an element of \code{ggd:::kinds} or an index number of
                     \code{ggd:::kinds}, or a \code{\link[ggd]{GGD}} object, or a \code{NA}.

                     If a character string is indicated as an element,
                     the string matches only one element of \code{ggd:::kinds} of
                     the first element along with the index order of
                     \code{ggd:::kinds.match.order} which is
                     \code{ c(1L, 4L, 2L, 3L, 7L, 5L, 6L, 10L, 8L, 9L, 13L, 11L, 12L,
                              16L, 14L, 15L)}.

                     The order of \code{ggd:::kinds.match.order} is designed,
                     while \code{ggd:::kinds} is ordered by intuitive degrees of freedom,
                     for practical purposes so that \code{"Mean-Differed Sigma-Differed"}
                     model, which has more degrees of freedom than the others of
                     the same type, can be matched first.

                     If an element is \code{NA}, \code{NA_integer_} will be returned for it.}

\item{undef.err}{A logical;
If \code{TRUE}, an error occur if a not-\code{NA} element of \code{objs}
does not match any element of \code{ggd:::kinds}.
If \code{FALSE}, \code{NA_integer_} will be returned for it.}
}
\value{
The vector of index numbers (integers) which represent the kinds of distributions.
         If \code{NULL} is given for \code{objs}, \code{NULL} will be returned.
}
\description{
Gets index numbers which represent the kinds of distributions.
}
\examples{
 ggd.kind.index( ggd:::kinds )
 ggd.kind.index( 1:16 )
 ggd.kind.index( 6 )
 ggd.kind.index( "3-Mean-Differed Sigma-Differed Vertical Gradational Distribution" )

 ggd.kind.index( "Horizontal" )              ## 7, not 5 nor a vector of 5:7.
 ggd.kind.index( "Mean-Eq.*Vertical" )       ## 9, not 12 nor a vector of c( 9, 12 ).
 ggd.kind.index( "3-Mean-Eq.*Vertical" )     ## 12

 ## If undef.err = FALSE (default),
 ## NA will be returned if the specified distribution is not supported.
 ggd.kind.index( "Poisson Distribution" )    ## NA

\dontrun{
 ## If undef.err = TRUE is specified, an error will occur.
 ggd.kind.index( "Poisson", undef.err = TRUE )

 ## Error: 'Poisson' does not match any character strings of kinds.}

 a <- GGD$new()
 ggd.kind.index( a )

 ## ggd.nls.freq.all returns a list of GGD objects to $obj.
 df <- data.frame( x = seq( -2, 2, 0.2 ),
                   freq = c(     57,    277,   1002,   3178,   9646,  22109, 42723,
                              80646, 117625, 139181, 162319, 150870, 109947, 78736,
                              46616,  21058,   9211,   3466,    976,    260,    61 ) )
 result <- ggd.nls.freq.all( df, not.use.nls = TRUE, start.level = 2 )
 ggd.kind.index( result$obj )
}
\seealso{
\code{\link[ggd]{ggd.kind}}
}
