% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{ggd.mix.type.for.kind.index}
\alias{ggd.mix.type.for.kind.index}
\title{Get mix.type according to kind.index}
\usage{
ggd.mix.type.for.kind.index(kind.index)
}
\arguments{
\item{kind.index}{A vector of index numbers of the kinds of distributions.
Values in \code{1:length( ggd:::kinds )} and \code{NA} are
allowed.}
}
\value{
A vector of integers of the mix.type which accepts the indicated
         kind of distribution.
}
\description{
Gets a vector of integers indicating \code{mix.type}s
which are appropriate to the indicated destributions by index numbers.
}
\examples{
 ggd.mix.type.for.kind.index( 1:7 )
}
\seealso{
\code{\link[ggd]{ggd.mix.type.for}}
}
