% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{ggd.start.template}
\alias{ggd.start.template}
\title{Template for start of nls}
\usage{
ggd.start.template(kind)
}
\arguments{
\item{kind}{A character string or a numeric value or a \code{\link[ggd]{GGD}} object
                 which indicates the kind of distribution model for which the template will
                 be output. The length should be 1.

                 The matching method of \code{kind} follows that of
                 \code{objs} argument of \code{\link[ggd]{ggd.kind.index}}.
                 If it matches to two or more distribution models,
                 this function returns only one template for the first matched model.}
}
\value{
A list containing components any of
         \item{mean}{
                 The start value for the mean value common to all normal distributions
                 of the components. 0 is preset.}
         \item{mean.i}{
                 The start value for the mean value of i-th normal distribution.
                 0 is preset.}
         \item{mean.i.j}{
                 The start value for the mean value of i,j-th normal distribution
                 in the 2x2 components. 0 is preset.}
         \item{sqrt.sd}{
                 The start value for the square root of the standard deviation common to
                 all normal distributions of the components. 1 is preset.}
         \item{sqrt.sd.i}{
                 The start value for the square root of the standard deviation of
                 i-th normal distribution. 1 is preset.}
         \item{sqrt.sd.i.j}{
                 The start value for the square root of the standard deviation of
                 i,j-th normal distribution in the 2x2 components. 1 is preset.}

         If \code{kind} represents an unsupported distribution model,
         \code{NULL} will be returned.
}
\description{
Gets a template list for \code{start} argument of \code{\link[ggd]{nls.freq}}
or for an element of \code{start} argument of \code{\link[ggd]{ggd.nls.freq.all}}.
This function can output only one template at a time.
So if you want to obtain two or more templates,
you should call this function one by one for each one.
}
\examples{
 ## preparing
 x <- seq( -2, 2, 0.2 )
 freq <- c( 1517,  2292,  2513,  2763,  3724,  4046,  4713,
            7947, 10997, 11824, 11133,  7868,  4692,  4103,
            3698,  2740,  2549,  2284,  1499,  1147,   918 )

 ## set the start parameters
 start <- ggd.start.template( 14 )
 start   ## check the parameters for the start of ggd:::kinds[14]

 start$mean.1.1 <- -0.671
 start$mean.1.2 <- -0.198
 start$mean.2.1 <-  0.293
 start$mean.2.2 <- -0.198
 start$sqrt.sd <- sqrt( 0.640 )  ## sqrt.sd is the square root of the standard deviation.

 ## try ggd.nls.freq
 ggd.nls.freq( data.frame( x, freq ), start = start, kind = 14 )$obj
}
\seealso{
\code{\link[ggd]{nls.freq}}, \code{\link[ggd]{ggd.nls.freq.all}},
         \code{\link[ggd]{ggd.init.start}}
}
