% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{initialize}
\alias{initialize}
\alias{new}
\alias{\S4method{new}{GGD}}
\alias{\S4method{initialize}{GGD}}
\title{Constructor}
\usage{
\S4method{new}{GGD}()
}
\description{
Generates a \code{\link[ggd]{GGD}} object.
The fields are initialized as a standard normal distribution model \eqn{\mathcal{N}(0, 1)}
with \code{mix.type = 2}, which represents a horizontal gradational distribution;
a basic model of the GGD.
To indicate components for a new \code{\link[ggd]{GGD}} object,
use \code{\link[ggd]{ggd.set.cmp}} or connect \code{\link[ggd]{set.cmp}} method as examples.
There are some other object generators:
\code{\link[ggd]{ggd.nls.freq}}, \code{\link[ggd]{ggd.nls.freq.all}} and
\code{\link[ggd]{ggd.trace.q}}.
}
\examples{
 a <- GGD$new()
 a
 rm( a )
 a <- GGD$new()$set.cmp( data.frame( mean = c( 0, 0 ), sd = c( 1.108, 0.703 ) ) )
 a
}
\seealso{
\code{\link[ggd]{ggd.set.cmp}},
         \code{\link[ggd]{ggd.nls.freq}}, \code{\link[ggd]{ggd.nls.freq.all}},
         \code{\link[ggd]{ggd.trace.q}}
}
