% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.default.custom.d}
\alias{is.default.custom.d}
\alias{\S4method{is.default.custom.d}{GGD}}
\title{Judge if custom.d is default}
\usage{
\S4method{is.default.custom.d}{GGD}()
}
\value{
\code{TRUE} if the function in \code{custom.d} field is
         identical with \code{ggd:::default.custom.d}, otherwise \code{FALSE}.
}
\description{
Checks if the function in \code{custom.d} field is
\code{\link[base]{identical}} with \code{ggd:::default.custom.d}.
}
\examples{
 a <- GGD$new()
 a$is.default.custom.d() ## TRUE

 a$set.cmp(
     this.cmp = data.frame( mean = c( -0.5, 0, 0.5 ), sd = c( 0.5, 0.75, 1 ) ),
     this.kind = "Custom",
     this.custom.d = function(x, cmp) (dnorm(x, cmp$mean[1], cmp$sd[1]) +
                                       dnorm(x, cmp$mean[2], cmp$sd[2]) +
                                       dnorm(x, cmp$mean[3], cmp$sd[3])) / 3 )
 a$is.default.custom.d() ## FALSE
}
\seealso{
\code{\link[ggd]{is.default.custom.p}}
}
