% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.default.custom.p}
\alias{is.default.custom.p}
\alias{\S4method{is.default.custom.p}{GGD}}
\title{Judge if custom.p is default}
\usage{
\S4method{is.default.custom.p}{GGD}()
}
\value{
\code{TRUE} if the character string of the \code{srcref} attribute
         of the function set in \code{custom.p} field
         is same as \code{ggd:::default.custom.p.text},
         otherwise \code{FALSE}.
}
\description{
Checks if the function in \code{custom.p} field can be considered the default
\code{custom.p} function.
This function converts the \code{srcref} attribute of the function in \code{custom.p} field
into a character string and compare it  with \code{ggd:default.custom.p.text}.
}
\examples{
 a <- GGD$new()
 a$is.default.custom.p() ## TRUE

 a$set.cmp(
     this.cmp = data.frame( mean = c( -0.5, 0, 0.5 ), sd = c( 0.5, 0.75, 1 ) ),
     this.kind = "Custom",
     this.custom.d = function(x, cmp) (dnorm(x, cmp$mean[1], cmp$sd[1]) +
                                       dnorm(x, cmp$mean[2], cmp$sd[2]) +
                                       dnorm(x, cmp$mean[3], cmp$sd[3])) / 3,
     this.custom.p = function(x, cmp) (pnorm(x, cmp$mean[1], cmp$sd[1]) +
                                       pnorm(x, cmp$mean[2], cmp$sd[2]) +
                                       pnorm(x, cmp$mean[3], cmp$sd[3])) / 3 )
 a$is.default.custom.p() ## FALSE
}
\seealso{
\code{\link[ggd]{is.default.custom.d}}
}
