% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.eq.mean}
\alias{is.eq.mean}
\alias{\S4method{is.eq.mean}{GGD}}
\title{Judge if all mean values are equal}
\usage{
\S4method{is.eq.mean}{GGD}()
}
\value{
\code{TRUE} if all \code{mean} values in \code{cmp} field are equal,
         otherwise \code{FALSE}.
}
\description{
Checks if the mean values of all normal distributions of components are equal.
The equality is determined by the \code{==} operator.
}
\examples{
 a <- GGD$new()
 a$kind          ## Normal Distribution
 a$is.eq.mean()  ## TRUE

 a$set.cmp( data.frame( mean = c( 0, 1e+8 ), sd = rep( 1.5, 2 ) ) )
 a$kind          ## Mean-Differed Sigma-Equaled Horizontal Gradational Distribution
 a$is.eq.mean()  ## FALSE

 a <- GGD$new()
 a$set.cmp( data.frame( mean = rep( 0.5, 2 ), sd = c( 1, 1 + 1e+8 ) ) )
 a$kind          ## Mean-Equaled Sigma-Differed Horizontal Gradational Distribution
 a$is.eq.mean()  ## TRUE
}
