% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.eq.sd}
\alias{is.eq.sd}
\alias{is.eq.sigma}
\alias{\S4method{is.eq.sd}{GGD}}
\alias{\S4method{is.eq.sigma}{GGD}}
\title{Judge if all standard deviations are equal}
\usage{
\S4method{is.eq.sd}{GGD}()

\S4method{is.eq.sigma}{GGD}()
}
\value{
\code{TRUE} if \code{sd} (= sigma) values in \code{cmp} field are equal,
         otherwise \code{FALSE}.
}
\description{
Checks if the standard deviations of all normal distributions of components are equal.
}
\examples{
 a <- GGD$new()
 a$kind          ## Normal Distribution
 a$is.eq.sd()    ## TRUE
 a$is.eq.sigma() ## TRUE (This method is just a synonym of is.eq.sd)

 a$trace.q(
     data.frame( x = c( -1.92, -0.20, 0.20, 1.92 ), p = c( 0.1, 0.4, 0.6, 0.9 ) ),
     grad = "v2" )
 a$kind          ## 2-Mean-Differed Sigma-Differed Vertical Gradational Distribution
 a$is.eq.sd()    ## FALSE
 a$is.eq.sigma() ## FALSE

 a$trace.q(
     data.frame( x = c( -0.67, 0, 0.53 ), p = c( 0.25, 0.5, 0.70 ) ),
     this.mix.type = 2, eq.sd = TRUE )
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )
 a$is.eq.sd()    ## TRUE
 a$is.eq.sigma() ## TRUE
}
