% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{is.hv}
\alias{is.hv}
\alias{\S4method{is.hv}{GGD}}
\title{Judge if horizontal-vertical gradation}
\usage{
\S4method{is.hv}{GGD}(strict = FALSE)
}
\arguments{
\item{strict}{If \code{TRUE}, this function returns \code{TRUE} only if the model is
                     precisely a horizontal-vertical gradational distribution and
                     it cannot be represented by any other simpler models.

                     If \code{FALSE}, this function also returns \code{TRUE} for
                     a normal distribution, a horizontal gradational distribution
                     and a 2-component vertical gradational distribution.}
}
\value{
\code{TRUE} if the object shows a kind of horizontal-vertical gradation of
         4 (2x2) normal distributions, otherwise \code{FALSE}.
}
\description{
Referring \code{mix.type} and \code{cmp} field, checks if the distribution is constructed as
a horizontal-vertical gradation of 4 (2x2) normal distributions.
}
\examples{
 a <- GGD$new()
 a$kind; a$cmp
 a$is.hv()       ## TRUE
 a$is.hv( TRUE ) ## FALSE

 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ),
                        sd = c( 1.1, 0.8, 1.2, 0.7 ) ), this.mix.type = 4 )
 a$is.hv()       ## TRUE
 a$is.hv( TRUE ) ## TRUE

 a$set.cmp( data.frame( mean = c( 0, 0, 1, 1 ),
                        sd = c( 1.1, 1.1, 1.2, 1.2 ) ), this.mix.type = 4 )
 a$is.hv()       ## TRUE
 a$is.hv( TRUE ) ## FALSE -- this is a horizontal gradational distribution

 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ),
                        sd = c( 1.1, 0.8, 1.1, 0.8 ) ), this.mix.type = 4 )
 a$is.hv()       ## TRUE
 a$is.hv( TRUE ) ## FALSE -- this is a vertical gradational distribution

 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ),
                        sd = c( 1.1, 0.8, 1.2, 0.8 ) ), this.mix.type = 4 )
 a$is.hv()       ## TRUE
 a$is.hv( TRUE ) ## TRUE -- this is NOT a vertical gradational distribution
}
