% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.normal}
\alias{is.normal}
\alias{\S4method{is.normal}{GGD}}
\title{Judge if normal distribution}
\usage{
\S4method{is.normal}{GGD}()
}
\value{
\code{TRUE} if the object shows a normal distribution, otherwise \code{FALSE}.
}
\description{
Referring only \code{cmp} field, checks if the distribution is a normal distribution.
Note, this function does not check \code{kind} and \code{kind.index} fields.
}
\details{
For \code{"Customized Distribution"}, this function returns always \code{FALSE}
even if the object has only 1 component and \code{custom.d} shows a normal distribution.
}
\examples{
 a <- GGD$new()
 a$is.normal()   ## TRUE

 a$set.cmp( data.frame( mean = c( 0, 0 ), sd = c( 1.1, 1.1 ) ), this.mix.type = 2 )
 a$is.normal()   ## TRUE

 a$trace.q(
     data.frame( x = c( -1.92, -0.20, 0.20, 1.92 ), p = c( 0.1, 0.4, 0.6, 0.9 ) ),
     grad = "v2" )
 a$kind; a$cmp
 a$is.normal()   ## FALSE
}
