% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{is.v2}
\alias{is.v2}
\alias{\S4method{is.v2}{GGD}}
\title{Judge if vertical gradation of 2 normal distributions}
\usage{
\S4method{is.v2}{GGD}(strict = FALSE)
}
\arguments{
\item{strict}{If \code{TRUE}, this function returns \code{TRUE} only if the model
                 is a vertical gradation with 2 different normal distributions.

                 If \code{FALSE}, this function also returns \code{TRUE} when the model
                 is a normal distribution.}
}
\value{
\code{TRUE} if the object shows a kind of vertical gradation of
         2 normal distributions, otherwise \code{FALSE}.
}
\description{
Referring \code{mix.type} and \code{cmp} field, checks if the distribution is constructed as
a vertical gradation of 2 normal distributions.
Even when \code{cmp} field has 3 rows with \code{mix.type = 3}, if the components of
the left-tail side and the right-tail side are the same, this function returns \code{TRUE},
because such a distribution can be reduced to 2-component model.
Also, if it is \code{mix.type = 4}
and essentially a vertical gradation of 2 normal distributions,
this function returns \code{TRUE}.
Note, this function does not check \code{kind} and \code{kind.index} fields.
}
\examples{
 a <- GGD$new()
 a$kind; a$cmp
 a$is.v2()       ## TRUE
 a$is.v2( TRUE ) ## FALSE

 a$trace.q(
     data.frame( x = c( -1.92, -0.20, 0.20, 1.92 ), p = c( 0.1, 0.4, 0.6, 0.9 ) ),
     grad = "v2" )
 a$kind; a$cmp
 a$is.v2( TRUE ) ## TRUE

 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ),
                        sd = c( 1.1, 0.8, 1.1, 0.8 ) ), this.mix.type = 4 )
 a$is.v2( TRUE ) ## TRUE -- this is a vertical gradational distribution

 a$set.cmp( data.frame( mean = c( 0, 0 ), sd = c( 1, 1 ) ), this.mix.type = 3 )
 a$is.v2()       ## TRUE  -- this is a vertical gradational distribution
 a$is.v2( TRUE ) ## FALSE -- this is a normal distribution
}
