% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{mean.calc}
\alias{mean.calc}
\title{[Non-exported] Mean calculation}
\usage{
mean.calc(mix.type, means, sds)
}
\arguments{
\item{mix.type}{The value of \code{mix.type}.}

\item{means}{The vector of the mean values of the normal distributions
of the components.}

\item{sds}{The vector of the standard deviations of the normal distributions
of the components.}
}
\value{
The mean value of a gradational Gaussian distribution model.
}
\description{
Calculates the mean value of a GGD model.
This function does not use \code{\link[stats]{integrate}}, but \code{\link[stats]{pnorm}},
\code{\link[base]{sqrt}} and four arithmetic operations as needed.
}
