% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{mean.calc.t4.sub}
\alias{mean.calc.t4.sub}
\title{[Non-exported] Subfunction of the mean calculation for mix.type = 4}
\usage{
mean.calc.t4.sub(means, sds)
}
\arguments{
\item{means}{The mean values of two normal distributions
\eqn{\mathcal{N}_{i,1}, \mathcal{N}_{i,2}}.}

\item{sds}{The standard deviations of two normal distributions
\eqn{\mathcal{N}_{i,1}, \mathcal{N}_{i,2}}.}
}
\value{
The value of \eqn{\int_{-\infty}^{\infty} x \Psi_i(x) g_i(x) dx}.
}
\description{
A subfunction which is used for calculating the mean value where \code{mix.type = 4}.
It calculates \eqn{\int_{-\infty}^{\infty} x \Psi_i(x) g_i(x) dx},
where
\eqn{\Psi_i(x) = \Phi_{i,1}(x) - \dfrac{1}{\sqrt{2}} \Phi_{i,1}^*(x) +
                 \dfrac{1}{\sqrt{2}} \Phi_{i,2}^*(x)},
\eqn{g_i(x) = ( 1 - f_{i,1}(x) / f_{i,1}(\mu_{i,1}) ) f_{i,1}(x) +
              f_{i,2}(x)^2 / f_{i,2}(\mu_{i,2})}
\eqn{( i = 1, 2 )}.
This function does not use \code{\link[stats]{integrate}},
but \code{\link[stats]{pnorm}}, \code{\link[base]{sqrt}} and four arithmetic operations.
}
