% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmn.norm.R
\name{ms.norm.xp}
\alias{ms.norm.xp}
\title{Mean and standard deviation of a normal distribution}
\usage{
ms.norm.xp(x, p)
}
\arguments{
\item{x}{The x-coordinates of the quantiles. It must be a vector with 2 numerics.}

\item{p}{The probabilities for the quantiles.
In other word, the values of the cumulative distribution function of
a normal distribution for the x. It must be also a vector with 2 numerics.}
}
\value{
A list containing components
         \item{mean}{
                 The mean value of the normal distribution.}
         \item{sd}{
                 The standard deviation of the normal distribution.}
}
\description{
Calculates the mean and standard deviation of the normal distribution
satisfying the given two quantiles (x-probability points).
This is the body of \code{ggd.trace.q(kind = "Normal Distribution")}.
}
\examples{
 ms.norm.xp( x = c( -1, 1 ), p = pnorm( c( -1, 1 ), 0, 1 ) ) # list( mean = 0, sd = 1 )
 ms.norm.xp( x = c(  0, 1 ), p = pnorm( c(  0, 1 ), 0, 2 ) ) # list( mean = 0, sd = 2 )
 ms.norm.xp( x = c( -2, 1 ), p = c( 0.3, 0.7 ) ) # list( mean = 0.5, sd = 2.86 ) (about)
}
