% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.file.R
\name{read.csv}
\alias{read.csv}
\alias{ggd.read.csv}
\alias{\S4method{read.csv}{GGD}}
\title{Read a composition}
\usage{
ggd.read.csv(file)

\S4method{read.csv}{GGD}(file)
}
\arguments{
\item{file}{The name of the CSV file.
                 The \code{file} can be a readable text-mode \link[base]{connection}.

         The \code{file} should be a CSV file with 3 columns.
         The columns are 1:row header, 2:mean value, and 3:standard deviation
         of the component. The columns cannot be reordered.
         The \code{file} must have a header row consisting of the value of
         \code{mix.type}, the character string of \code{"mean"}, and \code{"sd"}.
         The second and subsequent rows should have the row name (such as \code{"nd.1"}),
         the mean value, and standard deviation of each component.

         If the row header is \code{"custom.d"} or \code{"custom.p"},
         the character string in the 2nd column of that row is used as the function
         for \code{custom.d} or \code{custom.p} field.
         In this case, the 3rd column is not used.

         For more information about the properties of this argument,
         see \code{file} argument of \link[utils]{read.table}.}
}
\value{
The generated \code{\link[ggd]{GGD}} object (invisible for \code{GGD} method).

         For \code{GGD} method: If failed to read, the object will be cleared.
}
\description{
Reads a CSV file recorded the composition of a \code{\link[ggd]{GGD}} object.
Then \code{ggd.read.csv} function generates a \code{\link[ggd]{GGD}} object,
and \code{read.csv} method sets the fields according to the composition.
}
\details{
\subsection{Reading empty data}{
     If the file contains a cleared object (a file with only the header row),
     the values of \code{kind}, \code{kind.index}, and \code{mix.type} fields
     will be \code{NA},
     and then \code{median}, \code{mean}, and other numeric fields will be \code{NaN}.
}
\subsection{Illegal format file}{
     If the format of the file was illegal and a \code{\link[ggd]{GGD}} object cannot
     be constructed, an error occurs and the \code{\link[ggd]{GGD}} object is cleared.

     If \code{mix.type} value stored in the file does not match the values for \code{cmp},
     an error may occur, or it may happen to work.
     However, it is not recommended to create such a non-conforming file anyway.
}
}
\examples{
 csvfile <- tempfile( fileext = ".csv" )
 a <- ggd.set.cmp( data.frame( mean = c( -1.739, 1.195 ), sd = c( 1.175, 1.831 ) ),
                   grad = "h" )
 a$write.csv( csvfile )

 ggd.read.csv( csvfile )
 b <- ggd.read.csv( csvfile )
 b$mix.type == a$mix.type
 b$cmp == a$cmp

 a$clear()
 a$read.csv( csvfile )
 a
 a$mix.type == b$mix.type
 a$cmp == b$cmp

 unlink( csvfile )
}
\seealso{
\code{\link[ggd]{write.csv}}
}
