% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.apply.R
\name{round.cmp}
\alias{round.cmp}
\alias{signif.cmp}
\alias{\S4method{round.cmp}{GGD}}
\alias{\S4method{signif.cmp}{GGD}}
\title{Rounding values in cmp}
\usage{
\S4method{round.cmp}{GGD}(dg.mean = 0, dg.sd = dg.mean)

\S4method{signif.cmp}{GGD}(dg.mean = 6, dg.sd = dg.mean)
}
\arguments{
\item{dg.mean}{A vector of integers indicating the numbers of
decimal places (\code{round.cmp}) or
significant digits (\code{signif.cmp}) for \code{mean} column.
It may work with any length. If the length is longer than
number of components or is not a divisor of number of components,
the extra length will be cut off.
If the length is \code{0} (including \code{NULL}),
mean values will not be rounded.}

\item{dg.sd}{A vector of integers similar to \code{dg.mean}, for \code{sd} column.
If the length is \code{0} (including \code{NULL}),
standard deviations will not be rounded.
See \code{\link[base]{round}} for more information.}
}
\value{
The processed \code{\link[ggd]{GGD}} object itself.
}
\description{
Rounds the values of all elements in the \code{cmp} field and adjusts other fields
accordingly.
\code{\link[base]{round}} and \code{\link[base]{signif}} functions can apply.
These functions does not change the number of rows in \code{cmp} field,
no matter what results are obtained.
If an error occurs, these functions do not clear the object, but retains the values of
all fields.
}
\examples{
 a <- ggd.set.cmp( data.frame( mean = c( 112.7865, 113.1647 ),
                                 sd = c( 7.936188, 11.092769 ) ) )
 a$cmp; a$kind; c( a$mean, a$sd )

 a$round.cmp( 1, 2 )
 a$cmp; a$kind; c( a$mean, a$sd )

 a$signif.cmp( 2 )
 a$cmp; a$kind; c( a$mean, a$sd )

 ## For the IEC 60559 standard
 a <- ggd.set.cmp( data.frame( mean = c( 9.5, 10.5 ), sd = c( 0.35, 0.45 ) ) )
 a$round.cmp( 0, 1 ) ## mean will be c( 10, 10 ). sd depends on OS.
 a$cmp

 ## It will be a normal distribution with 2 rows in cmp.
 a <- ggd.set.cmp( data.frame( mean = c( 1.12, 1.06 ), sd = c( 1.99, 2.07 ) ) )
 a$round.cmp()
 a$cmp; a$kind; c( a$mean, a$sd )

 ## Only mean will be rounded.
 a <- ggd.set.cmp( data.frame( mean = c( -1.341789, 0.982471 ),
                                 sd = c( 1.897149, 2.412378 ) ) )
 a$signif.cmp( dg.mean = 2, dg.sd = NULL )
 a$cmp
}
\seealso{
\code{\link[ggd]{apply.cmp}}, \code{\link[base]{round}}
}
