% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{separate.data.quarter}
\alias{separate.data.quarter}
\title{[Non-exported] Quartering data}
\usage{
separate.data.quarter(x, data, data.mean)
}
\arguments{
\item{x}{A vector of x-coordinates arranged in ascending order
and non-duplicated.}

\item{data}{A vector of numeric values of the data with respect to x-coordinates.}

\item{data.mean}{Mean value of \code{data}.}
}
\value{
A list containing components
         \item{x}{
                 The vectors of x-coordinates of the 4 groups.}
         \item{data}{
                 The vectors of data of the 4 groups.}
         \item{[x|data].lower}{
                 The vector of x-coordinates or data integrated the lower-side 2 groups.}
         \item{[x|data].upper}{
                 The vector of x-coordinates or data integrated the upper-side 2 groups.}
         \item{[x|data].outer}{
                 The vector of x-coordinates or data integrated the outer-side 2 groups.}
         \item{[x|data].inner}{
                 The vector of x-coordinates or data integrated the inner-side 2 groups.}
}
\description{
Divides numeric data into 4 groups as:
at first, it divides the data into two groups by the mean of the data,
and next, divides each data into two groups so that the number of elements are equal to
each other.
If there is an x-coordinate that is exactly equal to the mean,
put that element in both the 2nd and 3rd part.
If the first dividing makes a group with odd number of elements, put the element at
the dividing point in both the lower and upper x-coordinate groups after second dividing.
}
