% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmn.norm.R
\name{sqnorm}
\alias{sqnorm}
\title{Sigma-unit distance from the mean of a normal distribution}
\usage{
sqnorm(p)
}
\arguments{
\item{p}{A vector of the probabilities of the quantiles.}
}
\value{
A vector of sigma-unit distances
         from the mean to the x-coordinates of the quantiles.
}
\description{
Calculates how many sigmas are the distance from the mean of a normal distribution
to the x-coordinate of the quantile which has the indicated probability.
The mean and standard deviation are not needed for this function
because the sigma-unit distance does not depend on them.
}
\examples{
 sqnorm( 0.5 )               ## 0
 sqnorm( pnorm( -2, 0, 1 ) ) ## -2
 sqnorm( seq( 0, 1, 0.1 ) )  ## increases from -Inf to Inf
}
