% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.trace.q.R
\name{v2.crossover}
\alias{v2.crossover}
\title{[Non-exported] Crossover-tracing (3 quantiles, vertical gradation of 2 normal distributions)}
\usage{
v2.crossover(qt, control)
}
\arguments{
\item{qt}{A data frame; the quantiles to be traced,
If the median is specified, median must be included.}

\item{control}{The \code{control} option for \code{\link[nleqslv]{nleqslv}}.}
}
\value{
A list containing components
         \item{cmp}{
                 The data frame for \code{cmp} field.}
         \item{result}{
                 The output of \code{\link[nleqslv]{nleqslv}}.}
}
\description{
Gets the data frame for \code{cmp} field where \code{mix.type = 3} and tracing 3 quantiles
with mean-differed 2 normal distributions of the components by crossover-tracing.
}
\details{
\subsection{Crossover-tracing}{
     Let cumulative distribution functions of 2 mean-differed normal distributions
     crossover at one of quantiles (except for median),
     and seek the mean values and standard deviations as tracing the other 2 quantiles.
 }
}
