% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{write}
\alias{write}
\alias{write.ggd}
\alias{\S4method{write}{GGD}}
\title{Write the composition of a object to a file}
\usage{
write.ggd(obj, file = "")

\S4method{write}{GGD}(file = "")
}
\arguments{
\item{obj}{The \code{\link[ggd]{GGD}} object to be saved.}

\item{file}{The name of the file or a \link[base]{connection} for writing
the composition of the object. \code{""} indicates output to the console.}
}
\value{
An invisible NULL.
}
\description{
Writes the composition of a \code{\link[ggd]{GGD}} object as a CSV file.
Mean values and standard deviations of the components are recorded to a maximum length of
the 22nd decimal place.
}
\examples{
 a <- ggd.set.cmp( data.frame( mean = c( 0.223, 0.219 ), sd = c( 2.265, 2.176 ) ),
                   vgrad.2 = TRUE )
 a$mix.type
 a$cmp
 a$write()
 write.ggd( a )
}
\seealso{
\code{\link[ggd]{read}}
}
