% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{adjust.kind.index}
\alias{adjust.kind.index}
\alias{\S4method{adjust.kind.index}{GGD}}
\title{Adjust kind and kind.index fields}
\usage{
\S4method{adjust.kind.index}{GGD}()
}
\value{
The adjusted \code{\link[ggd]{GGD}} object itself (invisible).
}
\description{
Sets \code{kind.index} and \code{kind} fields according to
\code{mix.type} and \code{cmp} fields.
Normally, users of this class don't need to call this method directly.
}
