% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.mean.sd.R
\name{calc.mean}
\alias{calc.mean}
\title{[Non-exported] Mean calculation}
\usage{
calc.mean(mix.type, means, sds)
}
\arguments{
\item{mix.type}{The value of \code{mix.type}. It allows from \code{1} to \code{4}.}

\item{means}{The vector of the mean values of the normal distributions
of the components.}

\item{sds}{The vector of the standard deviations of the normal distributions
of the components.}
}
\value{
The mean value of the distribution model.
}
\description{
Calculates the mean value of the distribution model.
}
\details{
This function does not use \code{\link[stats]{integrate}} or other tolerance-requiring
processes. Therefore, the calculation speed and accuracy will be expected in some degree.
However, since \code{\link[base]{sqrt}(pi)} is used where \code{mix.type} is \code{2}
or greater, and \code{\link[stats]{pnorm}} is also used where \code{mix_type = 4},
it is expected that there will be calculation errors due to their implementation,
in addition to digit losses during arithmetic operations.
}
