% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{calc.v}
\alias{calc.v}
\title{[Non-exported] Variance calculation}
\usage{
calc.v(
  mix.type,
  means,
  sds,
  mean = calc.mean(mix.type, means, sds),
  symmetric = FALSE,
  get.lv = FALSE,
  get.uv = FALSE
)
}
\arguments{
\item{mix.type}{The value of \code{mix.type} of the \code{\link[ggd]{GGD}} class.}

\item{means}{The mean values of the normal distributions of the components.}

\item{sds}{The standard deviations of the normal distributions of the components.}

\item{mean}{The mean of the whole distribution.}

\item{symmetric}{If \code{TRUE}, the distribution model is treated as symmetric.
This option may help reduce calculation errors.}

\item{get.lv}{If \code{TRUE}, this function culculates the lower half variance.}

\item{get.uv}{If \code{TRUE}, this function culculates the upper half variance.

                     Only one of \code{get.lv = TRUE} and \code{get.uv = TRUE} are valid;
                     if both are \code{TRUE}, \code{get.lv} takes priority.
                     If both are \code{FALSE} (the default),
                     this function culculates the whole variance.}
}
\value{
The value of the whole/half variance.
}
\description{
Calculates the variance or the half variance of a GGD model.
This function does not use \code{\link[stats]{integrate}},
but \code{\link[stats]{dnorm}}, \code{\link[stats]{pnorm}}, \code{\link[base]{sqrt}}
and four arithmetic operations.
}
\details{
For \code{mix.type = 4}, the half variance option of
\code{get.lv = TRUE} or \code{get.uv = TRUE} does not work. The reason is how to calculate
\eqn{\int_{-\infty}^{\mu} f_{i,1}(x) \Phi_{i,2}(x) dx} without numerical integral function
is still unknown for us.
}
