% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{calc.v.sub.t4}
\alias{calc.v.sub.t4}
\title{[Non-exported] A sub-function for variance calculation for mix.type = 4}
\usage{
calc.v.sub.t4(means, sds)
}
\arguments{
\item{means}{The mean values of the normal distributions
\eqn{\mathcal{N}_{i,1}} and \eqn{\mathcal{N}_{i,2}}.}

\item{sds}{The standard deviations of the normal distributions
\eqn{\mathcal{N}_{i,1}} and \eqn{\mathcal{N}_{i,2}}.}
}
\value{
The value of \eqn{\int_{-\infty}^{\infty} x^2 \Psi_i(x) g_i(x) dx}.
}
\description{
A sub-function for calculating the variance of a GGD model where \code{mix.type = 4}.
This function calculates \eqn{\int_{-\infty}^{\infty} x^2 \Psi_i(x) g_i(x) dx},
where
\eqn{\Psi_i(x) = \Phi_{i,1}(x) - \dfrac{1}{\sqrt{2}} \Phi_{i,1}^*(x) +
                                 \dfrac{1}{\sqrt{2}} \Phi_{i,2}^*(x)},
\eqn{g_i(x) = ( 1 - f_{i,1}(x) / f_{i,1}(\mu_{i,1}) ) f_{i,1}(x) +
              f_{i,2}(x)^2 / f_{i,2}(\mu_{i,2})}
with fixing \eqn{i}. This function does not use \code{\link[stats]{integrate}},
but \code{\link[stats]{pnorm}}, \code{\link[base]{sqrt}} and four arithmetic operations.
}
