% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{cat.table}
\alias{cat.table}
\title{[Non-exported] Output a table to a file}
\usage{
cat.table(x, file = "", top.text = "", digits = 15)
}
\arguments{
\item{x}{The object to be written, should be a matrix or data frame.}

\item{file}{A \link[base]{connection},
or a character string naming the file to print to.
If \code{""} (the default), it prints to the standard output connection,
the console unless redirected by \link[base]{sink}.}

\item{top.text}{A character string to be output as the first element of the header.}

\item{digits}{The number of significant (see \link[base]{signif}) digits to print.
Valid values are integers from 1 to 22 with default 15.}
}
\value{
An invisible NULL.
}
\description{
Prints the argument \code{x} (it should be a data frame or matrix)
to a file or \link[base]{connection} with the CSV format.
In particular, this function does not use any kind of \link[base]{paste} function to output,
but only \link[base]{cat} function.
When outputting real numbers, the precision can be specified with \code{digits} argument.
}
\examples{
 ggd:::cat.table( data.frame( a = 1:6 * pi, b = pi^(1:6) ), top.text = "PI", digits = 10 )
}
