% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{get.cmp.tex}
\alias{get.cmp.tex}
\title{[Non-exported] TeX format for mean and standard deviation}
\usage{
get.cmp.tex(obj, format.num)
}
\arguments{
\item{obj}{A \code{\link[ggd]{GGD}} object.}

\item{format.num}{A function to format each number of value of a parameter;
mean value and standard deviation. See \code{\link[ggd]{tex}}.}
}
\value{
TeX-formatted texts to represent the mean values and standard deviations of
         the normal dintributions of the components.
}
\description{
Gets the TeX-formatted texts to represent the mean values and standard deviations of
the normal dintributions of the components.
}
