% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{ggd.get.t3.cmp}
\alias{ggd.get.t3.cmp}
\title{Get cmp field for mix.type = 3}
\usage{
ggd.get.t3.cmp(means, sds, grad)
}
\arguments{
\item{means}{A vector of mean values of the 3 normal distributions of the components.}

\item{sds}{A vector of standard deviations of the 3 normal distributions of
the components.}

\item{grad}{A character string indicating the method of gradation.
If "v3", the number of components is forced to be 3.}
}
\value{
The data frame for \code{cmp} field.
}
\description{
Gets a data frame for \code{cmp} field for \code{mix.type = 3}.
Each of \code{means} and \code{sds} arguments must has 3 elements in order of
[1] left-tail side, [2] top side, [3] right-tail side.
}
\examples{
 means <- c( 0.2, 0, 0.2 ); sds <- c( 1.2, 0.9, 1.2 )
 ggd.get.t3.cmp( means, sds )                    # 2 components
 ggd.get.t3.cmp( means, sds, grad = "v3" )       # 3 components

 means <- rep( 0.5, 3 ); sds <- rep( 2.5, 3 )
 ggd.get.t3.cmp( means, sds )                    # 2 components even if normal distribution
}
