% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{ggd.mix.type.for}
\alias{ggd.mix.type.for}
\title{Get mix.type according to grad}
\usage{
ggd.mix.type.for(
  grad = c("default", "normal", "h", "v", "v2", "v3", "hv"),
  kind = NULL,
  mix.type = NULL
)
}
\arguments{
\item{grad}{A character string indicating the method of gradation.

                     \code{"h"} for horizontal, \code{"v"} for vertical,
                     and \code{"hv"} for horizontal-vertical.
                     The number after \code{"v"} is the number of components.
                     Numberless \code{"v"} is an alias for \code{"v2"}.

                     \code{"normal"} is for a normal distribution.
                     \code{"default"} is for depending on other arguments.}

\item{kind}{A character string or a numeric value or a \code{\link[ggd]{GGD}} object
                     which indicates the kind of distribution model.
                     This argument is enabled when the length of \code{mix.type} argument
                     is \code{0}, and \code{grad} is \code{"default"}.

                     The matching method of this argument follows that of elements of
                     the \code{objs} argument of the \link[ggd]{ggd.kind.index}.

                     If \code{NA} is indicated, \code{NA_integer_} will be returned.
                     If indicated character string pattern or index number does not match
                     to any element of \code{ggd:::kind}, an error will occur.}

\item{mix.type}{Default value which is returned if \code{grad} is \code{"default"}.
Although return value will become in integer class,
this function does not check if the value of this argument is valid.}
}
\value{
An integer of \code{mix.type} value appropriate for a distribution model
         represented by \code{grad} value or the default value indicated other arguments.
         If the indicated default value is \code{NULL},
         \code{integer(0)} is returned when \code{grad} is \code{"default"}.

         Although the length of \code{grad} argument must be 1 (or 0 as the default),
         but lengths of other arguments are not checked in this function consciously.
         So if \code{grad} is \code{"default"}, a vector of 2 or more length can be returned.
         That means, the length of other arguments or the return value must be checked
         by the caller of this function if necessary.
}
\description{
Gets the value for \code{mix.type} field of \code{GGD} object which is appropriate to
given \code{grad} value.
If \code{grad = "default"}, the return value will be the value of \code{mix.type} argument
or the value appropriate for \code{kind} argument.
}
\examples{
 ggd.mix.type.for( grad = "normal" )                 ## 0
 ggd.mix.type.for( grad = "h" )                      ## 2
 ggd.mix.type.for( grad = "v2" )                     ## 3
 ggd.mix.type.for( kind = "Normal" )                 ## 0
 ggd.mix.type.for( grad = "default", kind = NA )     ## NA
 ggd.mix.type.for( grad = "def", mix.type = NULL )   ## integer(0)
 ggd.mix.type.for( grad = "def", mix.type = -1 )     ## -1 (invalid value as mix.type)
}
\seealso{
\code{\link[ggd]{ggd.mix.type.for.kind.index}}, \code{\link[ggd]{ggd.ncmp.for}}
}
