% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{ggd.nls.freq.all}
\alias{ggd.nls.freq.all}
\title{Approximating a frequency distribution with all of supported models}
\usage{
ggd.nls.freq.all(
  data,
  x = "x",
  freq = "freq",
  total = NULL,
  start.level = 100,
  start = NULL,
  control = list(maxiter = 300, warnOnly = TRUE),
  not.use.nls = FALSE,
  cor.method = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame which represents the frequency distribution.
                 It must contain at least 2 numeric columns for \code{x} and \code{freq}.
                 See \code{\link[ggd]{nls.freq}} for more information.

                 Column names and indexes for \code{x} and \code{freq} are flexible.
                 You can specify them with next two arguments.}

\item{x}{The column name or index number for the column of x-coordinates
in \code{data}.}

\item{freq}{The column name or index number for the column of frequencies
in \code{data}.}

\item{total}{Total value of the frequencies.
See \code{\link[ggd]{nls.freq}} for more information.}

\item{start.level}{A numeric value of integer in from \code{0} to \code{3} or \code{100}
with default \code{100};
the level at which to guess the initial \code{start} parameters
of \code{\link[stats]{nls}}.
See \code{\link[ggd]{nls.freq}} for more information.}

\item{start}{A \bold{list of lists} with the length of 16
                 for each of the \code{start} arguments for \code{\link[stats]{nls}}
                 as initial values.
                 Each element (a list) will give to the \code{start} argument of
                 \code{\link[stats]{nls}} one by one.

                 For initial values for which \code{NULL} is indicated as the list
                 in the \code{start} argument, internally computed initial values
                 depending on \code{start.level} are used.

                 As an auxiliary tool for making a list of 16 lists,
                 you can use \code{\link[ggd]{ggd.init.start}} function
                 which returns a list of 16 \code{NULL}s.
                 And to get a template list for each list, you can use
                 \code{\link[ggd]{ggd.start.template}} function.

                 In addition, \code{\link[ggd]{ggd.kind}} and
                 \code{\link[ggd]{ggd.kind.index}} functions may help you whether
                 each index number of \code{start} represents what kind of distribution.
                 See "Examples" for usages of these tools.}

\item{control}{The \code{control} argument for \code{\link[stats]{nls}}.
See \code{\link[stats]{nls.control}} for more information.}

\item{not.use.nls}{A logical.
                 If \code{TRUE}, this function does not use \code{\link[stats]{nls}}
                 and it outputs objects having the initial values in the \code{cmp} field
                 as the results.
                 If \code{FALSE}, this function uses \code{\link[stats]{nls}}.

                 This argument works when \code{start.level} argument is
                 other than \code{100}.
                 A warning will occur if \code{TRUE} when \code{start.level} is \code{100}.

                 You can use \code{not.use.nls = TRUE} to check the initial values
                 when an error has occurred at this function.}

\item{cor.method}{The \code{method} argument for \code{\link[stats]{cor}}.
It represents the correlation coefficient method.
If \code{NULL}, it uses the default method of \code{\link[stats]{cor}}.
See \code{\link[stats]{cor}} for more information.}

\item{...}{Each argument for \code{\link[stats]{nls}} can be indicated.
See "Arguments" of \code{\link[stats]{nls}} for more information.}
}
\value{
A list containing components (invisible for \code{GGD} method)
         \item{best}{
                 The \code{\link[ggd]{GGD}} object which has got
                 the highest correlation coefficient.
                 That is, the most approximate to the given frequency distribution
                 in all of the supported distribution models.
                 If there are some models which have got the same highest correlation
                 coefficient, the object with the earlier \code{kind.index} is
                 given priority.}

         \item{best.cor}{
                 The correlation coefficient of \code{best} for the given
                 frequency distribution.}

         \item{obj}{
                 The list of 16 \code{\link[ggd]{GGD}} objects
                 ordered by \code{kind.index}; the index number in \code{ggd:::kinds}.
                 If an error has occurred, the element will be a cleared object.}

         \item{cor}{
                 The vector of the correlation coefficient of
                 between the result of each model and the frequency distribution.
                 \code{NA} will be given for an error case or an extremely bad result.}

         \item{detail}{
                 The list of 16 elements of outputs of \code{\link[ggd]{nls.freq}}.
                 Normally, each element is a list of the output of
                 \code{\link[ggd]{nls.freq}},
                 but if an error has occurred, the element will be an error condition.
                 See "Value" of \code{\link[ggd]{nls.freq}} for more information.}
}
\description{
Approximates the given frequency distribution with all of distribution models available in
this package except for \code{"Custom Distribution"} (the number of models is 16),
and compare their accuracies.
The accuracy is checked by the correlation coefficients with the frequency distribution
computed by \code{\link[stats]{cor}}.
}
\details{
The output lists are ordered by \code{ggd:::kinds} as:
\enumerate{
     \item   Normal Distribution
     \item   Mean of Mean-Differed Sigma-Equaled 2 Normal Distributions
     \item   Mean of Mean-Equaled Sigma-Differed 2 Normal Distributions
     \item   Mean of Mean-Differed Sigma-Differed 2 Normal Distributions
     \item   Mean-Differed Sigma-Equaled Horizontal Gradational Distribution
     \item   Mean-Equaled Sigma-Differed Horizontal Gradational Distribution
     \item   Mean-Differed Sigma-Differed Horizontal Gradational Distribution
     \item   2-Mean-Differed Sigma-Equaled Vertical Gradational Distribution
     \item   2-Mean-Equaled Sigma-Differed Vertical Gradational Distribution
     \item   2-Mean-Differed Sigma-Differed Vertical Gradational Distribution
     \item   3-Mean-Differed Sigma-Equaled Vertical Gradational Distribution
     \item   3-Mean-Equaled Sigma-Differed Vertical Gradational Distribution
     \item   3-Mean-Differed Sigma-Differed Vertical Gradational Distribution
     \item   Mean-Differed Sigma-Equaled Horizontal-Vertical Gradational Distribution
     \item   Mean-Equaled Sigma-Differed Horizontal-Vertical Gradational Distribution
     \item   Mean-Differed Sigma-Differed Horizontal-Vertical Gradational Distribution
}
Each index numbers of above list are
also used for \code{kind.index} field of \code{\link[ggd]{GGD}} class.

This function generates 16 \code{\link[ggd]{GGD}} objects and
calls \code{\link[ggd]{nls.freq}} method 16 times.
By default, \code{\link[ggd]{nls.freq}} is called with \code{warnOnly = TRUE},
so \code{\link[ggd]{nls.freq}} does not generate errors, but generates warnings often.
When a warning occur, this function generates another warning like
"\code{Warning for kind = xx :}" to inform which \code{kind.index} gets a poor result
(poor, but may be accurate enough).
So when one warning has occurred, two warnings will occur eventually.

If you indicate \code{warnOnly = FALSE} in \code{control} argument
and overwrite \code{warnOnly} option, \code{\link[ggd]{nls.freq}} can generate errors.
If an error occurs in one of \code{\link[ggd]{nls.freq}} processes,
this function throws messages like "\code{Error for kind = xx :}" and "\code{Error in ...}"
instead of throwing error and does not stop,
then tries other \code{\link[ggd]{nls.freq}} processes.
For the result of error-occurred \code{kind.index}, a cleared \code{\link[ggd]{GGD}} object
will be got as an element of \code{obj} (see "Value").
}
\examples{
 ## Preparing.
 plot.freq.and.d <- function( obj, x, freq )
 {
     xlim <- c( min( x ), max( x ) )
     ylim <- c( 0, max( ggd:::get.d.freq( x, freq ) ) * 1.2 )
     plot( x, ggd:::get.d.freq( x, freq ), xlim = xlim, ylim = ylim, xlab = "", ylab = "" )
     par( new = TRUE )
     plot( seq( min( x ), max( x ), 0.01 ), obj$d( seq( min( x ), max( x ), 0.01 ) ),
           type = "l", xlim = xlim, ylim = ylim )
 }

 df <- data.frame(
             x = seq( -2, 2, 0.2 ),
             freq = c( 1517,  2292,  2513,  2763,  3724,  4046,  4713,
                       7947, 10997, 11824, 11133,  7868,  4692,  4103,
                       3698,  2740,  2549,  2284,  1499,  1147,   918 ) )

 ## Try ggd.nls.freq.all.
 ## Here, "start.level = 1" is specified to get less-than-ideal results
 ## to explain how to use "start" argument.
 result <- ggd.nls.freq.all( df, start.level = 1 )

 ## Show the results.
 result$cor
 result$best.cor
 result$best

 ## Check that the value of cor for kind = 14 is very low.
 result$cor[[14]]

 ## Let's try to increase the value of cor by changing the initial values.
 ##
 ## There is an easy and good way to solve this problem,
 ## that is to remove "start.level = 1".
 ## But here, we use "start" argument with remaining "start.level" for explaining.
 ##
 ## First, to see what kind = 14 is, display the kind.
 result$obj[[14]]$kind

 ## Also, using ggd.kind, you can get the character string for the index.
 ggd.kind( 14 )

 ## Inversely, using ggd.kind.index, you can get index for each kind.
 ggd.kind.index( "Mean-Differed Sigma-Equaled Horizontal-Vertical" )

 ## Show the cmp field of kind = 14 and plots.
 result$obj[[14]]$cmp
 plot.freq.and.d( result$obj[[14]], df$x, df$freq )

 ## Now, for the initial values, we are going to use the result of
 ## kind = "2-Mean-Differed Sigma-Equaled Vertical Gradational Distribution".
 ggd.kind.index( "2-Mean-Differed Sigma-Equaled Vertical" )  ## 8
 result$obj[[8]]$cmp

 ## Display the parameters for the initial values for kind = 14.
 ggd.start.template( 14 )

 ## Set the initial values for kind = 14.
 start.list <- ggd.init.start()
 start.list[[14]] <- ggd.start.template( 14 )
 start.list[[14]]$mean.1.1 <- result$obj[[8]]$cmp$mean[1]
 start.list[[14]]$mean.1.2 <- result$obj[[8]]$cmp$mean[2]
 start.list[[14]]$mean.2.1 <- result$obj[[8]]$cmp$mean[1]
 start.list[[14]]$mean.2.2 <- result$obj[[8]]$cmp$mean[2]
 start.list[[14]]$sqrt.sd <- sqrt( result$obj[[8]]$cmp$sd[1] )   ## Set sqrt for SD.

 ## Retry ggd.nls.freq.all.
 result <- ggd.nls.freq.all( df, start.level = 1, start = start.list )
 result$cor
 result$cor[[14]]
 result$obj[[14]]$cmp
 plot.freq.and.d( result$obj[[14]], df$x, df$freq )
}
\seealso{
\code{\link[ggd]{nls.freq}}, \code{\link[stats]{cor}},
         \code{\link[ggd]{ggd.init.start}}, \code{\link[ggd]{ggd.start.template}}
}
