% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{ggd.start.template}
\alias{ggd.start.template}
\title{Template for start of nls}
\usage{
ggd.start.template(target, n.cmp = 0)
}
\arguments{
\item{target}{A variable that identifies the kind of distribution to which
the template applies. The length should be 1.
The type of the variable is valid for \code{\link[ggd]{GGD}} object,
the string of an element of \code{ggd:::kinds}, or its index number.}

\item{n.cmp}{Number of components for the custom distribution.

                     If \code{target} indicates \code{"Custom Distribution"} or \code{17},
                     a positive integer must be indicated for this argument.

                     If a \code{\link[ggd]{GGD}} object is indicated for \code{target},
                     the number of rows of \code{cmp} field is used instead
                     when \code{n.cmp} is \code{0} (the default).
                     When \code{n.cmp} is a positive integer, the value of \code{n.cmp}
                     takes priority.}
}
\value{
A list containing components any of
         \item{mean}{
                 The start value for mean values common to all normal distributions
                 of the components. 0 is preset.}
         \item{mean.i}{
                 The start value for the mean value of i-th normal distribution.
                 0 is preset.}
         \item{mean.i.j}{
                 The start value for mean value of i,j-th normal distribution
                 in the 2x2 components. 0 is preset.}
         \item{sqrt.sd}{
                 The sqrt of the start value for standard deviations common to
                 all normal distributions of the components. 1 is preset.}
         \item{sqrt.sd.i}{
                 The sqrt of the start values for the standard deviation of
                 i-th normal distribution. 1 is preset.}
         \item{sqrt.sd.i.j}{
                 The sqrt of the start values for the standard deviation of
                 i,j-th normal distribution in the 2x2 components. 1 is preset.}

         If \code{target} represents unsupported distribution kind, \code{NULL} is returned.
}
\description{
Gets the template for start list for \code{\link[stats]{nls}}
when you want to provide your own \code{start} parameters for \code{\link[ggd]{nls.freq}}
or \code{\link[ggd]{ggd.nls.freq.all}}.
This function can output only one template list at once.
So if you want to get several templates for different distribution kinds,
call this function one by one for each distribution kind.
}
\examples{
 ## preparing
 x <- seq( -2, 2, 0.2 )
 freq <- c( 1517,  2292,  2513,  2763,  3724,  4046,  4713,
            7947, 10997, 11824, 11133,  7868,  4692,  4103,
            3698,  2740,  2549,  2284,  1499,  1147,   918 )

 ## set the start parameters
 start <- ggd.start.template( 14 )
 start   ## check the parameters for the start of ggd:::kinds[14]

 start$mean.1.1 <- -0.671
 start$mean.1.2 <- -0.198
 start$mean.2.1 <-  0.293
 start$mean.2.2 <- -0.198
 start$sqrt.sd <- sqrt( 0.640 )  ## sqrt.sd is the square root of the standard deviation.

 ## try ggd.nls.freq
 ggd.nls.freq( data.frame( x, freq ), start = start, kind = 14 )$obj
}
\seealso{
\code{\link[ggd]{nls.freq}}, \code{\link[ggd]{ggd.nls.freq.all}},
         \code{\link[ggd]{ggd.init.start}}
}
