% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{is.eq.mean}
\alias{is.eq.mean}
\alias{\S4method{is.eq.mean}{GGD}}
\title{Judge if all mean values are equal}
\usage{
\S4method{is.eq.mean}{GGD}()
}
\value{
\code{TRUE} if all \code{mean} values in \code{cmp} field are equal,
         otherwise \code{FALSE}.
}
\description{
Checks if the mean values of all normal distributions of components are equal.
}
\examples{
 a <- GGD$new()
 a$kind          ## Normal Distribution
 a$is.eq.mean()  ## TRUE

 a$trace.q(
     data.frame( x = c( -1.92, -0.20, 0.20, 1.92 ), p = c( 0.1, 0.4, 0.6, 0.9 ) ),
     grad = "v2" )
 a$kind          ## 2-Mean-Differed Sigma-Differed Vertical Gradational Distribution
 a$is.eq.mean()  ## FALSE

 a <- GGD$new()
 a$set.cmp( data.frame( mean = rep( 0, 3 ), sd = c( 1.2, 0.5, 1.3 ) ) )
 a$kind          ## 3-Mean-Equaled Sigma-Differed Vertical Gradational Distribution
 a$is.eq.mean()  ## TRUE
}
