% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.eq.sd}
\alias{is.eq.sd}
\alias{is.eq.sigma}
\alias{\S4method{is.eq.sd}{GGD}}
\alias{\S4method{is.eq.sigma}{GGD}}
\title{Judge if all standard deviations are equal}
\usage{
\S4method{is.eq.sd}{GGD}()

\S4method{is.eq.sigma}{GGD}()
}
\value{
\code{TRUE} if \code{sd} (= sigma) values in \code{cmp} field are equal,
         otherwise \code{FALSE}.
}
\description{
Checks if the standard deviations of all normal distributions of components are equal.
The equality is determined by the '\code{==}' operator.
}
\examples{
 a <- GGD$new()
 a$kind          ## Normal Distribution
 a$is.eq.sd(); a$is.eq.sigma()   ## TRUE; TRUE (is.eq.sigma is just a synonym of is.eq.sd)

 a$set.cmp( data.frame( mean = rep( 0.5, 2 ), sd = c( 1, 1 + 1e+8 ) ) )
 a$kind          ## Mean-Equaled Sigma-Differed Horizontal Gradational Distribution
 a$is.eq.sd(); a$is.eq.sigma()   ## FALSE; FALSE

 a$set.cmp( data.frame( mean = c( 0, 1e+8 ), sd = rep( 1.5, 2 ) ) )
 a$kind          ## Mean-Differed Sigma-Equaled Horizontal Gradational Distribution
 a$is.eq.sd(); a$is.eq.sigma()   ## TRUE; TRUE
}
