% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.normal}
\alias{is.normal}
\alias{\S4method{is.normal}{GGD}}
\title{Judge if normal}
\usage{
\S4method{is.normal}{GGD}()
}
\value{
\code{TRUE} if the object shows a normal distribution, otherwise \code{FALSE}.
}
\description{
Referring \code{mix.type} and \code{cmp} field, checks if the distribution is essentially
a normal distribution.
Note, this function does not check \code{kind} and \code{kind.index} fields.
}
\details{
For \code{"Custom Distribution"}, this function returns always \code{FALSE}
even if the object has only 1 component and \code{custom.d} shows a normal distribution.
}
\examples{
 a <- GGD$new()
 a$is.normal()   ## TRUE

 a$set.cmp( data.frame( mean = c( 0, 0 ), sd = c( 1.1, 1.1 ) ), this.mix.type = 2 )
 a$is.normal()   ## TRUE

 a$set.cmp( data.frame( mean = c( 0, 0 ), sd = c( 1.1, 0.9 ) ), grad = "v2" )
 a$kind          ## 2-Mean-Equaled Sigma-Differed Vertical Gradational Distribution
 a$is.normal()   ## FALSE
}
