% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.symmetric}
\alias{is.symmetric}
\alias{\S4method{is.symmetric}{GGD}}
\title{Judge if symmetric}
\usage{
\S4method{is.symmetric}{GGD}()
}
\value{
\code{TRUE} if the probability density function of the distribution model is
         symmetric about the mean, otherwise \code{FALSE}.
}
\description{
Referring \code{mix.type} and \code{cmp} field,
checks if the probability density function is symmetric about the mean.
}
\details{
This method judges that the distribution model is symmetric if either:
\enumerate{
     \item   a normal distribution.
     \item   a mean of 2 normal distributions or a horizontal gradational distribution
             where the two components have the same standard deviation.
     \item   a mean of 2 normal distributions or a 2-component vertical gradational
             distribution where the two components have the same mean value.
     \item   a 3-component vertical gradational distribution
             where the both tail-side components have the same standard deviation,
             and the mean value of the top-side component equals the mean of the mean values
             of the both tail-side components.
     \item   a horizontal-vertical gradational distribution which has
             two vertical gradational distributions in the components
             which are symmetrical to each other about the mean of the distribution.
}
}
\examples{
 a <- GGD$new()
 a$is.symmetric()    ## TRUE

 a$set.cmp( data.frame( mean = c( -0.35, 0.35 ), sd = c( 1, 1 ) ), grad = "h" )
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )
 a$is.symmetric()    ## TRUE

 a$set.cmp( data.frame( mean = c( 0, 0 ), sd = c( 1.1, 0.8 ) ), grad = "v2" )
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )
 a$is.symmetric()    ## TRUE

 a$adjust.cmp( grad = "hv" )
 a$is.symmetric()    ## TRUE

 a$set.cmp( data.frame( mean = c( -0.8, -0.2, 0.8, 0.2 ), sd = c( 1.2, 0.9, 1.2, 0.9 ) ) )
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )
 a$is.symmetric()    ## TRUE

 a$set.cmp( data.frame( mean = c( -0.8, -0.2, 0.2, 0.8 ), sd = c( 1.2, 0.9, 1.2, 0.9 ) ) )
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )
 a$is.symmetric()    ## FALSE
}
