% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{is.v3}
\alias{is.v3}
\alias{\S4method{is.v3}{GGD}}
\title{Judge if 3-component vertical}
\usage{
\S4method{is.v3}{GGD}(strict = FALSE)
}
\arguments{
\item{strict}{If \code{TRUE}, this function returns \code{TRUE} only if
                 the distribution is a vertical gradation with different left-tail-side and
                 right-tail-side components.
                 Even when \code{cmp} field has 3 rows with \code{mix.type = 3},
                 if the 1st and the 3rd components are the same, it returns \code{FALSE}.

                 If \code{FALSE}, this function also returns \code{TRUE} for
                 a normal distribution or a vertical gradation of 2 normal distributions.}
}
\value{
\code{TRUE} if the object represents a kind of vertical gradation of
         3 normal distributions, otherwise \code{FALSE}.
}
\description{
Referring \code{mix.type} and \code{cmp} field, checks if the distribution is essentially
a vertical gradation of 3 normal distributions.
Note! when \code{mix.type = 4},
if the right-top-side and left-top-side components are the same
and both tail-side components are different from each other,
this method returns \code{FALSE}.
Because it is not a vertical gradation of 3 normal distributions, but it contains
an element of the horizontal gradation of the 2 tail-side components.
Note, this function does not check \code{kind} and \code{kind.index} fields.
}
\examples{
 a <- GGD$new()
 a$kind; a$cmp
 a$is.v3()                   ## TRUE
 a$is.v3( strict = TRUE )    ## FALSE

 a$set.cmp( data.frame( mean = c( 0, 0, 0 ), sd = c( 1.1, 0.9, 1.2 ) ), grad = "v3" )
 a$is.v3()                   ## TRUE
 a$is.v3( strict = TRUE )    ## TRUE

 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ),
                        sd = c( 1.1, 0.8, 1.2, 0.8 ) ), this.mix.type = 4 )
 a$mix.type  ## 4
 a$is.v3()   ## FALSE -- see 'Note!' at Description
}
