% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{nls.freq.level.100}
\alias{nls.freq.level.100}
\title{[Non-exported] Loop nls for start.level = 100}
\usage{
nls.freq.level.100(
  data,
  total,
  kind,
  mix.type,
  grad,
  eq.mean,
  eq.sd,
  control,
  method,
  ...
)
}
\arguments{
\item{data}{A data frame which represents the frequency distribution.
It must contain 2 numeric columns named \code{x} and \code{freq}.}

\item{total}{Total value of the frequencies.}

\item{kind}{An object indicating the \code{kind} of \code{\link[ggd]{GGD}} object.}

\item{mix.type}{A numeric value to set into \code{mix.type} field of
the \code{\link[ggd]{GGD}} object as an integer.
You can specify a value in \code{0:4}.}

\item{grad}{A character string indicating the method of gradation.
See \code{\link[ggd]{nls.freq}} for more information.}

\item{eq.mean}{A logical. If \code{TRUE}, all of the mean values of of the components
are forced to be equal.}

\item{eq.sd}{A logical. If \code{TRUE}, all of the standard deviations of
the components are forced to be equal.}

\item{control}{The list for \code{control} argument of \code{\link[stats]{nls}}.
See \code{\link[stats]{nls.control}} for more information.}

\item{method}{The \code{method} argument for \code{\link[stats]{cor}},
which represents the correlation coefficient method.}

\item{...}{Each argument for \code{\link[stats]{nls}} can be indicated.
See "Arguments" of \code{\link[stats]{nls}} for more information.}
}
\value{
A list conforming the return value of \code{\link[ggd]{nls.freq}}.
}
\description{
Execute a loop of \code{\link[stats]{nls}} with changing the level of initial guessing.
This is the main process of \code{\link[ggd]{nls.freq}} with \code{start.level = 100}.
}
\seealso{
\code{\link[ggd]{nls.freq}}
}
