% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.nls.freq.R
\name{nls.freq}
\alias{nls.freq}
\alias{ggd.nls.freq}
\alias{\S4method{nls.freq}{GGD}}
\title{Approximating a frequency distribution}
\usage{
ggd.nls.freq(data, x = "x", freq = "freq", total = NULL,
         kind = NULL, mix.type = NULL,
         grad = c("default", "normal", "h", "v", "v2", "v3", "hv"),
         eq.mean = logical(), eq.sd = logical(),
         start.level = 100, start = NULL, control = list(),
         not.use.nls = FALSE, cor.method = NULL, ...)

\S4method{nls.freq}{GGD}(data, x = "x", freq = "freq", total = NULL,
         this.kind = NULL, this.mix.type = NULL,
         grad = c("default", "normal", "h", "v", "v2", "v3", "hv"),
         eq.mean = logical(), eq.sd = logical(),
         start.level = 100, start = NULL, control = list(),
         not.use.nls = FALSE, cor.method = NULL, ...)
}
\arguments{
\item{data}{A data frame which represents the frequency distribution.
                     It must contain at least 2 numeric columns for \code{x} and \code{freq}.

                     Column \bold{\code{x}} is for the x-coordinates.
                     Each value expected to be a numeric value which represents the cell of
                     the frequency distribution, the x-coordinate at the center of a cell
                     of the frequency distribution.
                     The values must be arranged in ascending order, and not be duplicated.

                     Column \bold{\code{freq}} is for the frequencies following \code{x}.
                     The values of frequencies must be positive.
                     Both integers and real numbers are allowed for the values.

                     Rows which contain \code{NA} or \code{NaN} for \code{x} or \code{freq}
                     are ignored. The number of rows should be large enough;
                     it is recommended that there are more than 8 valid rows.
                     At least, 3 valid rows must be contained.

                     Column names and column numbers for \code{x} and \code{freq}
                     are flexible. You can specify them with next two arguments.}

\item{x}{The column name or column number for x-coordinates in \code{data}.}

\item{freq}{The column name or column number for frequencies in \code{data}.}

\item{total}{Total value of the frequencies.

                     If \code{NULL} (the default),
                     the total of \code{freq}, i.e., \code{\link[base]{sum}(data[[freq]])}
                     (on \code{\link[stats]{complete.cases}} of \code{x} and \code{freq})
                     is used for it.}

\item{kind}{A character string or a numeric value or a \code{\link[ggd]{GGD}} object
                     which indicates the kind of distribution model for approximating
                     the frequency distribution.

                     The matching method of this argument follows that of elements of
                     the \code{objs} argument of the \code{\link[ggd]{ggd.kind.index}}.

                     This argument gives the conditions of
                     the value of \code{mix.type} field,
                     and of whether the mean values and standard deviations of the components
                     should be aligned to the same value.

                     Indicating \code{mix.type} argument or
                     indicating other than \code{"default"} for \code{grad} argument
                     or \code{TRUE}/\code{FALSE} for \code{eq.mean} or \code{eq.sd}
                     can overwrite the conditions of this argument.}

\item{mix.type}{A numeric value to set into \code{mix.type} field of
                     the \code{\link[ggd]{GGD}} object as an integer.
                     It should be an integer from \code{0} to \code{4} or \code{NULL}.

                     The type of the distribution model will be as:
                     \itemize{
                         \item 0: Normal distribution.
                         \item 1: Mean of 2 normal distributions.
                         \item 2: Horizontal gradation of 2 normal distributions.
                         \item 3: Vertical gradation of 2 or 3 normal distributions.
                                  The 2-component model has priority.
                         \item 4: Horizontal-Vertical gradation
                                  with 4 (2x2) normal distributions.
                     }

                     If other than \code{"default"} for \code{grad} argument is indicated,
                     this argument will be ignored.}

\item{grad}{A character string indicating the method of gradation.

                     \code{"h"} for horizontal, \code{"v"} for vertical,
                     and \code{"hv"} for horizontal-vertical.
                     The number after \code{"v"} is the number of components.
                     Numberless \code{"v"} is an alias for \code{"v2"}.

                     \code{"normal"} is for a normal distribution.
                     \code{"default"} is for depending on values of other arguments.

                     If other than \code{"default"} is indicated,
                     this function will create a \code{\link[ggd]{GGD}} object
                     according to this argument with ignoring \code{[this.]mix.type} argument
                     and overwriting the type indicated by \code{[this.]kind} argument.}

\item{eq.mean}{A logical. If \code{TRUE}, all of the mean values of of the components
                     are forced to be equal.

                     If \code{FALSE} or \code{logical(0)},
                     the mean values of the components can be different to each other,
                     and may be equal in very rare cases.

                     \code{TRUE} and \code{FALSE} can overwrite the condition indicated by
                     \code{kind} or \code{this.kind} argument.}

\item{eq.sd}{A logical. If \code{TRUE}, all of the standard deviations of
                     the components are forced to be equal.

                     If \code{FALSE} or \code{logical(0)},
                     the standard deviation of the components can be different to
                     each other, and may be equal in very rare cases.

                     If both \code{eq.mean} and \code{eq.sd} are \code{TRUE},
                     a normal distribution will be generated.

                     \code{TRUE} and \code{FALSE} can overwrite the condition indicated by
                     \code{kind} or \code{this.kind} argument.}

\item{start.level}{A numeric value of integer in from \code{0} to \code{3} or \code{100}
                     with default \code{100}; the level at which to guess the initial
                     \code{start} parameters of \code{\link[stats]{nls}}.

         Details for each level are as:
         \itemize{
             \item \code{0}:
                      The mean and the standard deviation of the frequency distribution
                      are used as initial values.
             \item \code{1}:
                      In addition to level \code{0}, if it is likely to be better guess,
                      it computes the mean values or standard deviations in ranges of
                      local x-coordinates where the effect of each of components is likely
                      to be heavy, and uses them as initial values.
             \item \code{2}:
                      In addition to level \code{0}, if it is likely to be better guess,
                      it uses the mean values or standard deviations of normal distributions
                      tracing two of quantiles which are generated with
                      the frequency distribution as initial values.
             \item \code{3}:
                      It uses the mean values and standard deviations of the components
                      of a \code{\link[ggd]{GGD}} object tracing some (2, 3 or 5) quantiles
                      which are generated with the frequency distribution as initial values.
                      If tracing fails, level \code{2} is used instead.
             \item \code{100}:
                      Try all of above levels and adopt the result with the highest
                      \code{\link[stats]{cor}} value.
         }

         The higher the level in the range of from \code{0} to \code{3},
         the more likely it is that the initial values will model
         the frequency distribution in closer,
         but the accuracy of the result may not along with the level.
         It is possible that \code{\link[stats]{nls}} will succeed at level \code{1}
         and fail at level \code{3} for the same data.}

\item{start}{A list of \code{start} argument for \code{\link[stats]{nls}}.

                     You can provide your own \code{start} for \code{\link[stats]{nls}},
                     the mean values
                     (names are like: \code{mean}, \code{mean.i} or \code{mean.i.j})
                     and the \bold{square root} of the standard deviations
                     (names are like: \code{sqrt.sd}, \code{sqrt.sd.i} or \code{sqrt.sd.i.j})
                     of the normal distributions of the components.

                     Depending on the kind of the distribution model,
                     the name of the parameters are different.
                     You can use \code{\link[ggd]{ggd.start.template}}
                     to get the template of the list and know the names of the parameters.

                     If a not-\code{NULL} list is indicated for this argument,
                     \code{start.level} argument is ignored.}

\item{control}{The list for \code{control} argument of \code{\link[stats]{nls}}.
See \code{\link[stats]{nls.control}} for more information.}

\item{not.use.nls}{A logical.
                     If \code{TRUE}, this function does not use \code{\link[stats]{nls}} and
                     it outputs an object having the initial values in the \code{cmp} field
                     as the result.
                     If \code{FALSE}, this function uses \code{\link[stats]{nls}}.

                     This argument works when \code{start.level} argument is
                     other than \code{100}. A warning will occur if \code{TRUE}
                     when \code{start.level} is \code{100}.

                     You can use \code{not.use.nls = TRUE} to check the initial values
                     when an error has occurred at this function.}

\item{cor.method}{The \code{method} argument for \code{\link[stats]{cor}}.
It represents the correlation coefficient method.
This argument is used only if \code{start.level = 100}.
If \code{NULL}, it uses the default method of \code{\link[stats]{cor}}.
See \code{\link[stats]{cor}} for more information.}

\item{...}{Each argument for \code{\link[stats]{nls}} can be indicated.
See "Arguments" of \code{\link[stats]{nls}} for more information.}

\item{this.kind}{A string or a numeric value or a \code{\link[ggd]{GGD}} object
                     which indicates the kind of distribution model for approximating
                     the frequency distribution.

                     This argument will work as same as \code{kind} argument
                     of the generator function (signature '\code{NULL}').

                     When this method is called without \code{this.kind} argument
                     or other conditions, it attempt to retain the value of
                     \code{mix.type} field as much as possible, but not the value of
                     \code{kind} field, i.e., the condition whether the mean value or
                     standard deviation of each component is aligned may not be retained.
                     If you want to retain these conditions as well,
                     indicate the object itself to \code{this.kind} argument like as
                     \code{obj$nls.freq(data, this.kind = obj)}.}

\item{this.mix.type}{A numeric value to set into \code{mix.type} field as an integer.
                         This argument will work as same as \code{mix.type} of
                         the generator function (signature '\code{NULL}').

                     If both of \code{this.kind} and \code{this.mix.type} are not given
                     and \code{grad} argument is \code{"default"},
                     the current value of \code{mix.type} field will be retained,
                     and number of components will also.
                     But furthermore, the object has been cleared,
                     the \code{mix.type} field will be set to \code{2}, the initial value.}
}
\value{
A list containing components (invisible for \code{GGD} method)
         \item{obj}{
                 Generated \code{\link[ggd]{GGD}} object which most (at least locally)
                 closely approximates the given frequency distribution.
                 If \code{\link[stats]{nls}} has failed, it will be a cleared object.
                 For \code{\link[ggd]{GGD}} method, the \code{\link[ggd]{GGD}} object itself.}
         \item{nls.out}{
                 The list of the output of \code{\link[stats]{nls}}.
                 If \code{\link[stats]{nls}} has not been used, \code{NULL} will be set.
                 See "Value" of \code{\link[stats]{nls}} for more information.}
         \item{start.level}{
                 The initial guessing level which is used actually to gain \code{obj}.
                 If \code{start.level = 100} is indicated,
                 the level for the best result will be set.
                 When \code{start.level = 3} is indicated
                 and if initial guessing has failed, \code{2} will be set.
                 If \code{start} argument (not-\code{NULL}) is indicated,
                 \code{NA} will be set.}
         \item{start}{
                 The used \code{start} argument for the initial values
                 for \code{\link[stats]{nls}}.}
         \item{start.obj}{
                 A \code{\link[ggd]{GGD}} object corresponding to the initial values.
                 That is, a \code{\link[ggd]{GGD}} object in which
                 values of the above \code{start} are set directly to the \code{cmp} field.}
         \item{cor}{
                 The vector of the correlation coefficient of between the result for
                 each level of initial guessing in from \code{0} to \code{3}
                 and the frequency distribution.
                 This component is given only if \code{start.level = 100}.}
         \item{errors}{
                 A list of information about errors occurred in \code{\link[stats]{nls}}.
                 This component is given only if \code{start.level = 100}.
                 Each element in the list contains:
                 \itemize{
                     \item level: The level of initial guessing when the error has occurred.
                     \item message: The error message.
                 }}
         \item{warnings}{
                 A list of information about warnings occurred in \code{\link[stats]{nls}}.
                 This component is given only if \code{start.level = 100}.
                 The composition of each element is as same as for \code{errors}.}

         For \code{GGD} method: If an error occur, the object will be cleared in most cases.
}
\description{
With the non-linear least squares (\code{\link[stats]{nls}}),
constructs a \code{\link[ggd]{GGD}} object which (locally) most closely approximates
the given frequency distribution. "Locally" means that if the start value is modified,
more closely approximating model may be generated.
The outliers of the frequency distribution will not be excluded in this function.
If necessary, outliers should be excluded by preprocessing.
}
\details{
\subsection{Why the standard deviations for "start" are square-rooted?}{
     You know a standard deviation must be a non-zero positive value.
     But if you use standard deviations directly in the formula for \code{\link[stats]{nls}},
     they will sometimes drop into negative values while the Gauss-Newton algorithm is running
     and the algorithm will fail, even if it can reach convergence if done better.

     So, to avoid such failures, we use square roots of standard deviations and
     take squares of them in the formula for \code{\link[stats]{nls}}.
 }
}
\examples{
 ## Preparing
 df <- data.frame( x      = seq( -2, 2, 0.2 ),
                   freq   = c( 1517,  2292,  2513,  2763,  3724,  4046,  4713,
                               7947, 10997, 11824, 11133,  7868,  4692,  4103,
                               3698,  2740,  2549,  2284,  1499,  1147,   918 ),
                   x.2    = seq( -20, 20, 2 ),
                   freq.2 = c( .000974, .003797, .008523, .023142, .045017, .081743, .120990,
                               .142527, .124627, .106294, .078625, .059378, .045690, .042958,
                               .035760, .030938, .015675, .012516, .008139, .005114, .003582 ) )

 plot.freq.and.d <- function( obj, x, freq )
 {
     xlim <- c( min( x ), max( x ) )
     ylim <- c( 0, max( ggd:::get.d.freq( x, freq ) ) * 1.2 )
     plot( x, ggd:::get.d.freq( x, freq ), xlim = xlim, ylim = ylim, xlab = "", ylab = "" )
     par( new = TRUE )
     plot( seq( min( x ), max( x ), 0.01 ), obj$d( seq( min( x ), max( x ), 0.01 ) ),
           type = "l", xlim = xlim, ylim = ylim )
 }

 ## Examples
 ggd.nls.freq( df, grad = "normal" )
 a <- ggd.nls.freq( df, grad = "no" )$obj
 plot.freq.and.d( a, df$x, df$freq )

 ## "start.level" can be omitted (and you should omit to gain better results),
 ## but is indicated here for processing speed.
 a$nls.freq( df, this.mix.type = 1, start.level = 2 )
 a
 plot.freq.and.d( a, df$x, df$freq )

 a$nls.freq( df, start.level = 2,
             this.kind = "2-Mean-Differed Sigma-Equaled Vertical Gradational Distribution" )
 a
 plot.freq.and.d( a, df$x, df$freq )

 ## overwriting "Sigma-Differed" after "kind = a"
 b <- ggd.nls.freq( df, kind = a, eq.sd = FALSE, start.level = 2 )
 b
 plot.freq.and.d( b$obj, df$x, df$freq )

 ## You can set start parameters if you want.
 start.list <- ggd.start.template( 14 )
 start.list

 start.list$mean.1.1 <- -0.671
 start.list$mean.1.2 <- -0.198
 start.list$mean.2.1 <- 0.293
 start.list$mean.2.2 <- -0.198
 start.list$sqrt.sd <- sqrt( 0.640 ) ## sqrt.sd is the sqrt of the standard deviation.

 ## "start.level" is ignored when you have indicated start parameters.
 a$nls.freq( df, this.kind = 14, start.level = 1, start = start.list )
 a
 plot.freq.and.d( a, df$x, df$freq )

 ## When you use a GGD object consecutively,
 ## the field values set to the object in the previous session are retained
 ## (if no error has occurred).
 a$nls.freq( df, grad = "hv", eq.mean = TRUE, start.level = 2 )
 a
 plot.freq.and.d( a, df$x, df$freq )

 a$nls.freq( df, eq.mean = FALSE, start.level = 2 )   ## grad = "hv" is retained.
 a
 plot.freq.and.d( a, df$x, df$freq )

 ## Using "x.2" for x and "freq.2" for freq.
 a <- ggd.nls.freq( df, x = "x.2", freq = "freq.2", start.level = 2 )$obj
 a   ## default value of mix.type is 2
 plot.freq.and.d( a, df$x.2, df$freq.2 )
}
\seealso{
\code{\link[stats]{nls}}, \code{\link[stats]{nls.control}},
         \code{\link[ggd]{ggd.nls.freq.all}}, \code{\link[ggd]{ggd.start.template}}
}
