% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.1.R
\name{q}
\alias{q}
\alias{\S4method{q}{GGD}}
\title{Quantile function}
\usage{
\S4method{q}{GGD}(prob, tol = .Machine$double.eps * 16)
}
\arguments{
\item{prob}{A vector of probabilities.}

\item{tol}{The tolerance level for the convergence criterion.}
}
\value{
A vector of the x-coordinates with the cumulative distribution function
         is equal to the given probabilities in the tolerance level.
}
\description{
Gets the x-coordinates with the cumulative distribution function is equal to
the given probabilities (with a tolerance level).
This method works like \code{\link[stats]{qnorm}} for a normal distribution.
}
\examples{
 a <- GGD$new()
 a$trace.q(
     data.frame( x = c( -0.67, 0, 0.53 ), p = c( 0.25, 0.5, 0.75 ) ),
     this.mix.type = 2 )
 a$q( c( 0.25, 0.5, 0.75 ) )
 plot( seq( 0, 1, 0.01 ), a$q( seq( 0, 1, 0.01 ) ), type = "l" )
}
\seealso{
\code{\link[ggd]{d}}, \code{\link[ggd]{p}}, \code{\link[ggd]{r}}
}
