% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.file.R
\name{read.csv}
\alias{read.csv}
\alias{ggd.read.csv}
\alias{\S4method{read.csv}{GGD}}
\title{Read a composition from a CSV file}
\usage{
ggd.read.csv(file)

\S4method{read.csv}{GGD}(file)
}
\arguments{
\item{file}{The name of the CSV file which the composition of
                     a \code{\link[ggd]{GGD}} object is to be read from.
                     The \code{file} can be a readable text-mode \link[base]{connection}.

                     The \code{file} must have a header consisting of
                     the value of \code{mix.type}, the character string of "mean"
                     and the character string of "sd".
                     The order of each column is not interchangeable.
                     The second and subsequent rows should be the name of the row
                     (such as \code{"nd.1"}, but not be cared)
                     and the mean value and standard deviation of the each components
                     of the \code{cmp} field.

                     For more information about the properties of this argument,
                     see \code{file} argument of \link[utils]{read.table}.}
}
\value{
The generated \code{\link[ggd]{GGD}} object (invisible for \code{GGD} method).

         For \code{GGD} method: If failed to read, the object will be cleared.
}
\description{
Reads a CSV file recorded the composition of a \code{\link[ggd]{GGD}} object
and generates a \code{\link[ggd]{GGD}} object.
}
\details{
\subsection{Reading empty data}{
     If the read file is one which is written with a cleared object,
     \code{kind}/\code{kind.index} and \code{mix.type} fields of the generated object
     will be \code{NA},
     and \code{median} and \code{mean} and other numeric fields will be \code{NaN}.
}
\subsection{Illegal format file}{
     If the format of the read file was illegal and a \code{\link[ggd]{GGD}} object cannot
     be constructed, an error occurs and the \code{\link[ggd]{GGD}} object is cleared.

     If the \code{mix.type} value stored in a file does not match the values for \code{cmp},
     an error may occur, or it may happen to work. It is possible to create such a file
     using \code{ggd:::}\code{\link[ggd]{cat.table}} function which is non-exported.
     However, it is not recommended to create such a non-conforming file anyway.
}
}
\examples{
 csvfile <- tempfile( fileext = ".csv" )
 a <- ggd.set.cmp( data.frame( mean = c( -1.739, 1.195 ), sd = c( 1.175, 1.831 ) ),
                   grad = "h" )
 a$write.csv( csvfile )

 ggd.read.csv( csvfile )
 b <- ggd.read.csv( csvfile )
 b$mix.type == a$mix.type
 b$cmp == a$cmp

 a$clear()
 a$read.csv( csvfile )
 a
 a$mix.type == b$mix.type
 a$cmp == b$cmp

 unlink( csvfile )
}
\seealso{
\code{\link[ggd]{write.csv}}
}
