% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.set.cmp.R
\name{set.cmp}
\alias{set.cmp}
\alias{ggd.set.cmp}
\alias{\S4method{set.cmp}{GGD}}
\title{Setting components}
\usage{
ggd.set.cmp(cmp, kind = NULL, mix.type = NULL,
         grad = c("default", "normal", "h", "v", "v2", "v3", "hv"))

\S4method{set.cmp}{GGD}(this.cmp = .self$cmp,
         this.kind = NULL, this.mix.type = NULL,
         grad = c("default", "normal", "h", "v", "v2", "v3", "hv"))
}
\arguments{
\item{cmp}{A data frame for \code{cmp} field.

                     It must have just 2 columns named "\code{mean}" and "\code{sd}",
                     and its rows must be less than or equals to 4.
                     Row names are not required.

                     The \code{mean} column represents the mean values of
                     the normal distributions of the components.

                     The \code{sd} column represents the standard deviations of
                     the normal distributions of the components.

                     Having no rows is allowed,
                     but indicating \code{NULL} or having no columns is not allowed.}

\item{kind}{A character string or a numeric value or a \code{\link[ggd]{GGD}} object
                     which indicates the kind of distribution model to be generated.

                     The matching method of \code{kind} follows that of
                     \code{objs} argument of \code{\link[ggd]{ggd.kind.index}}.
                     \code{NA} is allowed when \code{cmp} has no rows.

                     If \code{mix.type} argument or other than \code{"default"} for
                     \code{grad} argument is indicated, this argument will be ignored.}

\item{mix.type}{A numeric value represents how to mix the normal distributions.

                     The type of the distribution model and the number of rows in \code{cmp}
                     will be represented with \code{mix.type} value as follows:
                     \itemize{
                         \item \code{0} : Normal distribution. \code{cmp} has only 1 row.
                         \item \code{1} : Mean of 2 normal distributions.
                                          \code{cmp} has 2 rows.
                         \item \code{2} : Horizontal gradational distribution.
                                          \code{cmp} has 2 rows.
                         \item \code{3} : Vertical gradational distribution.
                                          \code{cmp} has 2 or 3 rows.
                         \item \code{4} : Horizontal-vertical gradational distribution.
                                          \code{cmp} has 4 rows.
                     }

                     If the number of rows in \code{cmp} argument is different from
                     the number shown above, \code{cmp} field will be
                     redundant/simplified to have the number of rows as above,
                     if possible. If not possible, an error will occur.

                     You can indicate \code{mix.type = NA} only if \code{cmp} has no rows.

                     If \code{grad} argument other than \code{"default"} is indicated,
                     this argument will be ignored.}

\item{grad}{A character string indicating the method of gradation.

                     \code{"h"} for horizontal, \code{"v"} for vertical,
                     and \code{"hv"} for horizontal-vertical.
                     The number after \code{"v"} is the number of components.
                     Numberless \code{"v"} is an alias for \code{"v2"}.

                     \code{"normal"} is for a normal distribution,
                     then also, \code{'grad = "no"'} can be read as "no gradation".

                     \code{"default"} is, if \code{kind} or \code{mix.type} argument
                     is given, follows it, otherwise it depends on the number of columns
                     in \code{cmp} argument. If the number of columns in cmp is \code{2},
                     the current \code{mix.type} is retained or horizontal (default) is used.}

\item{this.cmp}{A data frame for setting into \code{cmp} field.
It is equivalent to \code{cmp} argument for \code{ggd.set.cmp}.}

\item{this.kind}{A string or a numeric value or a \code{\link[ggd]{GGD}} object
which indicates the kind of distribution model to be constructed.
It is equivalent to \code{kind} argument for \code{ggd.set.cmp}.}

\item{this.mix.type}{A numeric value represents how to mix the normal distributions.
It is equivalent to \code{mix.type} argument for \code{ggd.set.cmp}.}
}
\value{
The \code{\link[ggd]{GGD}} object itself (invisible for \code{GGD} method).

         For \code{GGD} method: If an error occurs, each value of field will not be changed.
}
\description{
Generates a \code{\link[ggd]{GGD}} object according to indicated values of \code{cmp} field
(\code{ggd.set.cmp}), or sets the values in \code{cmp} field and all of other fields
accordingly (\code{set.cmp}).
Whenever you want to set values in \code{cmp} field, it is strongly recommended to use
this method.
}
\details{
\subsection{About "kind" and "mix.type"}{
     In this function,
     unlike \code{\link[ggd]{trace.q}} and \code{\link[ggd]{nls.freq}} methods,
     \code{[this.]kind} argument is only used to determine \code{mix.type} value,
     which shows how to mix the normal distributions of the components.
     That is, \code{[this.]kind} argument has no effect to align the mean values or
     standard deviations of the components to be equal.

     So, the character string indicated as \code{[this.]kind} argument may not match
     the new value of \code{[this.]kind} field.
     For example, if you indicate \code{[this.]kind = "Mean-Eq.*Horizontal"} and
     \code{[this.]cmp = data.frame(mean = c(0, 1), sd = c(0.8, 1.2))},
     the new \code{kind} field will be
     \code{"Mean-Differed Sigma-Differed Horizontal Gradational Distribution"},
     which is not matched with indicated regular expression.
     In such a case, a warning will occur.

     On the other hand, when a \code{\link[ggd]{GGD}} object is
     indicated as \code{[this.]kind}, no warning will occur if the new \code{kind} field
     is different from that of the indicated object.
     Indicated object is regarded as just for specifying \code{mix.type} value.

     It is not recommended but if you indicate not-\code{NULL} values for both of
     \code{[this.]kind} and \code{[this.]mix.type} at once,
     \code{[this.]kind} will be ignored.

     If both \code{[this.]kind} and \code{[this.]mix.type} arguments are \code{NULL}
     and \code{grad} argument is \code{"default"},
     the new value of \code{mix.type} field will be decided according to
     number of rows in \code{[this.]cmp} as:
     \itemize{
         \item \code{nrow([this.]cmp) = 0} : \code{mix.type} will be \code{NA}.
         \item \code{nrow([this.]cmp) = 1} : \code{mix.type} will be \code{0}.
         \item \code{nrow([this.]cmp) = 2} : If the current \code{mix.type} field
                                             is in \code{1:3}, it will be retained.
                                             Otherwise, \code{mix.type} will be \code{2}.
         \item \code{nrow([this.]cmp) = 3} : \code{mix.type} will be \code{3}.
         \item \code{nrow([this.]cmp) = 4} : \code{mix.type} will be \code{4}.
     }
 }
}
\examples{
 ## Normal Distribution
 a <- ggd.set.cmp( data.frame( mean = 0, sd = 1.5 ) )
 a$mix.type; a$cmp
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )

 ## "normal" for "grad" argument is allowed, though, it does little work in this function.
 a <- ggd.set.cmp( data.frame( mean = 1, sd = 2 ), grad = "normal" )
 a$mix.type; a$cmp

 ## Where the number of the cmp field rows is 2,
 ## it is recommended to indicate "grad" or "mix.type" or "kind" to avoid confusing.
 rm( a )
 a <- ggd.set.cmp( data.frame( mean = c( 0, 0 ), sd = c( 1.2, 0.8 ) ),
                   kind = "Mean of Mean-Equaled Sigma-Differed 2 Normal Distributions" )
 a$mix.type; a$cmp
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )

 ## Changing to mix.type = 2 : Horizontal Gradational Distribution
 a$set.cmp( a$cmp, this.mix.type = 2 )
 a$mix.type; a$cmp

 ## You can also write as:
 a$set.cmp( a$cmp, grad = "h" )
 a$mix.type; a$cmp
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )

 ## kind.index = 9 : 2-Mean-Equaled Sigma-Differed Vertical Gradational Distribution
 a$set.cmp( a$cmp, this.kind = 9 )
 a$mix.type; a$cmp
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )

 ## 3-Mean-Differed Sigma-Differed Vertical Gradational Distribution
 a$set.cmp( data.frame( mean = c( -0.5, 0, 0.5 ), sd = c( 1.2, 0.8, 1.2 ) ) )
 a$mix.type; a$cmp
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )

 ## Mean-Equaled Sigma-Differed Horizontal-Vertical Gradational Distribution
 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ), sd = c( 0.7, 0.5, 2.0, 1.5 ) ) )
 a$mix.type; a$cmp
 plot( seq( -3, 3, 0.01 ), a$d( seq( -3, 3, 0.01 ) ), type = "l" )

 ## If "cmp" field can be simplified, "mix.type" field may be set to lower type.
 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ), sd = c( 1, 0.7, 1, 0.7 ) ) )
 a$mix.type  ## 3 (not 4)
 a$cmp       ## with 2 rows

 ## If you want not to simplify "mix.type" and "cmp" fields,
 ## indicate "mix.type" / "this.mix.type" or one of other arguments for the condition.

 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ), sd = c( 1, 0.7, 1, 0.7 ) ),
            this.mix.type = 4 )
 a$mix.type  ## 4
 a$cmp       ## with 4 rows

 ## You can also write as:
 a$set.cmp( data.frame( mean = c( 0, 0, 0, 0 ), sd = c( 1, 0.7, 1, 0.7 ) ),
            grad = "hv" )
 a$mix.type; a$cmp

 ## The "kind" / "this.kind" argument can also indicate
 ## how to mix the normal distributions of the components.
 ##
 ## However, if the components indicated by "cmp" or "this.cmp" are inconsistent
 ## with the value of "kind" / "this.kind" argument,
 ## the "kind" and "kind.index" fields will be set to different values with a warning.
 ## For avoiding confusing, you should use "mix.type" / "this.mix.type" or "grad" instead.

 ## This sample will work with a warning.
 a$set.cmp(
     this.cmp = data.frame( mean = c( 0, 0, 0, 0 ), sd = c( 1, 0.7, 1, 0.7 ) ),
     this.kind = "Mean-Equaled Sigma-Differed Horizontal-Vertical" )
 a$mix.type  ## 4
 a$kind      ## "2-Mean-Equaled Sigma-Differed Vertical Gradational Distribution"
}
