% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGD.R
\name{sqnorm}
\alias{sqnorm}
\title{Sigma-unit distance from the mean of a normal distribution}
\usage{
sqnorm(p)
}
\arguments{
\item{p}{A vector of probabilities.}
}
\value{
A vector of sigma-unit distances from the mean to the quantiles for
         the probabilities.
}
\description{
Calculates how many sigmas are the distances from the mean of a normal distribution
to the values of the quantile function for the given probabilities.
The mean and standard deviation are not need for this function, because sigma-unit distances
do not depend on them.
}
\examples{
 sqnorm( 0.5 )   # 0
 sqnorm( pnorm( -2, 0, 1 ) ) # -2
 sqnorm( seq( 0, 1, 0.1 ) )  # increces from -Inf to Inf
}
