% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.trace.q.R
\name{v3.qt4.cmp}
\alias{v3.qt4.cmp}
\title{[Non-exported] Gets cmp field (4 quantiles, 3-component vertical gradational distribution)}
\usage{
v3.qt4.cmp(qt, eq.mean, eq.sd, control, retriable, grad)
}
\arguments{
\item{qt}{A data frame; the quantiles to be traced.}

\item{eq.mean}{A logical; the flag to be equal all of the mean values.}

\item{eq.sd}{A logical; the flag to be equal all of the standard deviations.}

\item{control}{The \code{control} option for \code{\link[nleqslv]{nleqslv}}.}

\item{retriable}{A logical; the flag if retrying is enable.}

\item{grad}{A character string indicating the method of gradation.
If \code{"v3"}, constructing with 3 components is enforcedly,
even if it is possible to construct with 2 components.}
}
\value{
A list containing components
         \item{cmp}{
                 The data frame for \code{cmp} field.}
         \item{result}{
                 The output of \code{\link[nleqslv]{nleqslv}}.}
}
\description{
Tracing 4 quantiles with 3 normal distributions, makes the data frame for \code{cmp} field
where \code{mix.type = 3}.
}
