% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd.file.R
\name{write.csv}
\alias{write.csv}
\alias{ggd.write.csv}
\alias{\S4method{write.csv}{GGD}}
\title{Write the composition}
\usage{
ggd.write.csv(obj, file = "")

\S4method{write.csv}{GGD}(file = "")
}
\arguments{
\item{obj}{The \code{\link[ggd]{GGD}} object to be saved.}

\item{file}{The name of the file or a \link[base]{connection} for writing
the composition of the object. \code{""} indicates output to the console.}
}
\value{
An invisible NULL.
}
\description{
Writes the composition of a \code{\link[ggd]{GGD}} object as a CSV file.
This function provides a simple way to export a \code{\link[ggd]{GGD}} object,
regardless of the package or R version. See "Details" for more information.
}
\details{
\subsection{Accuracy of saved data}{
     Mean values and standard deviations are recorded to a maximum length of
     the 22nd decimal place.
     The accuracy is sufficient to reconstruct the original object almost completely
     (at least the value of each field can be \code{TRUE} with \code{"=="})
     in most cases, and in most systems.
}
\subsection{For custom distribution}{
     Although it is possible to save an object of \code{"Customized Distribution"}
     (\code{mix.type = 5}), it is not always a good idea to use this method.
     This is because, \code{costom.d} is simply converted to a character string
     and saved in the last row of the CSV file, so it is not guaranteed that objects loaded
     with \code{\link[ggd]{read.csv}} will work well,
     particularly if \code{custom.d} calls your own functions.
     Instead, you can use \code{\link[base]{save}} or \code{\link[base]{save.image}}.
}
}
\examples{
 a <- ggd.set.cmp( data.frame( mean = c( 0.223, 0.219 ), sd = c( 2.265, 2.176 ) ),
                   grad = "v2" )
 a$mix.type
 a$cmp
 a$write.csv()
 ggd.write.csv( a )
}
\seealso{
\code{\link[ggd]{read.csv}}
}
